Method to aggregate peptides into proteins with the iterative approach.

Method to aggregate peptides into proteins with the iterative approach

aggIter(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)

Arguments

qPepData
A matrix of intensities of peptides.

X
An adjacency matrix

conditions
xxx

init.method
The method used to initialize the iterative algotirhm. Default is 'Sum'.

method
The method used for xxx. Default value is 'Mean'.

n
xxxx

Value

A matrix of protein intensities.

Author

Samuel Wieczorek

Examples

library(QFeatures)

#> Loading required package: MultiAssayExperiment

#> Loading required package: SummarizedExperiment

#> Warning: package ‘SummarizedExperiment’ was built under R version 4.0.2

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: ‘IRanges’

#> The following object is masked from ‘package:grDevices’:
#> 
#>     windows

#> Loading required package: GenomeInfoDb

#> Warning: package ‘GenomeInfoDb’ was built under R version 4.0.2

#> Loading required package: DelayedArray

#> Warning: package ‘DelayedArray’ was built under R version 4.0.2

#> Loading required package: Matrix

#> 
#> Attaching package: ‘Matrix’

#> The following object is masked from ‘package:S4Vectors’:
#> 
#>     expand

#> Loading required package: MatrixGenerics

#> Warning: package ‘MatrixGenerics’ was built under R version 4.0.2

#> Loading required package: matrixStats

#> Warning: package ‘matrixStats’ was built under R version 4.0.2

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> 
#> Attaching package: ‘MatrixGenerics’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars

#> The following object is masked from ‘package:Biobase’:
#> 
#>     rowMedians

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply, rowsum

#> 
#> Attaching package: ‘QFeatures’

#> The following object is masked from ‘package:DelayedArray’:
#> 
#>     sweep

#> The following object is masked from ‘package:base’:
#> 
#>     sweep

utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]], 'all'))
conditions <- colData(obj)$Condition
aggIter(assay(obj,2), X, conditions)

#>            [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#>   [1,] 24.28821 24.48162 24.32201 24.10576 24.28980 25.45394
#>   [2,] 24.23783 23.38166      NaN      NaN      NaN      NaN
#>   [3,]      NaN      NaN      NaN      NaN      NaN      NaN
#>   [4,] 22.50959 22.67980 22.93771 22.55190 22.71791 22.58516
#>   [5,] 26.09462 26.21429 26.16749 26.09004 26.01065 26.09375
#>   [6,] 24.17805 24.10837 23.74755 24.13438 24.88646 24.02547
#>   [7,] 21.92136 21.55936      NaN 22.78337 22.34627      NaN
#>   [8,] 23.19862 22.03234 23.18264 22.87017 22.59332 22.85274
#>   [9,] 27.73982 27.63242 27.89151 27.59328 27.68826 27.73286
#>  [10,] 22.01828 22.31921 22.13877 22.46914 22.61715 22.52920
#>  [11,]      NaN      NaN 21.70092      NaN      NaN      NaN
#>  [12,] 22.62053 23.07604 23.03971 23.12210 22.78965 23.08937
#>  [13,] 26.04675 26.16701 26.16127 26.09831 26.08242 26.03132
#>  [14,] 30.34763 30.31490 30.43516 30.38733 30.38177 30.35366
#>  [15,] 23.12366 23.25507 23.25249 23.32313 23.07801 23.35860
#>  [16,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [17,] 23.91031 23.95136 23.77967 24.12305 23.96455 23.86703
#>  [18,] 29.87954 29.89272 30.03522 30.40243 30.29909 30.42482
#>  [19,] 21.31163 21.27509 22.31722 20.93049 21.14582 21.36051
#>  [20,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [21,] 23.73469 23.78726 23.98555      NaN      NaN      NaN
#>  [22,]      NaN      NaN      NaN      NaN 22.17283      NaN
#>  [23,] 23.82978 24.91452      NaN      NaN 24.56651 24.36212
#>  [24,] 25.62327 25.72882 25.62374 25.57467 25.59867 25.66909
#>  [25,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [26,] 26.03183 26.00093 26.12761 26.00755 25.96299 25.92105
#>  [27,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [28,] 24.66239 24.58734 24.82125 24.67311 24.65929 24.70313
#>  [29,]      NaN      NaN      NaN 23.50252 23.54687 23.77957
#>  [30,] 24.21076 24.17064 25.04714 24.76779 26.41700 24.47977
#>  [31,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [32,] 23.66839 23.77806 23.64240 24.33470 24.31347 24.11012
#>  [33,] 24.93463 24.87037 24.71618 24.91884 24.78905 24.80464
#>  [34,] 24.09233 23.91424 24.00273 23.92488 24.03607 24.13312
#>  [35,] 23.12307 22.98801 22.78391 22.30048 23.16987 23.14247
#>  [36,] 24.76855 24.84460 25.07997 24.80371 24.84575 24.95180
#>  [37,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [38,] 25.29399 25.27184 25.40099 25.43472 25.36470 25.15994
#>  [39,] 25.91166 25.37806 26.37814 25.56976 25.30946 25.97235
#>  [40,] 25.52925 24.96531 25.45001 24.62304 23.68222 24.56691
#>  [41,] 25.44038 24.82527 24.95916 25.05887      NaN 25.89809
#>  [42,] 28.38187 28.62697 28.24436 28.71954 28.74114 28.69962
#>  [43,] 24.30904 24.16507 24.14857 24.55462 24.05606 23.84048
#>  [44,] 29.00603 28.83993 29.03708      NaN      NaN      NaN
#>  [45,] 22.74411 23.07084 23.43288 23.80863 23.67347 23.63998
#>  [46,] 22.10856 22.91929 22.77429      NaN 23.45237 22.30265
#>  [47,] 23.50736 23.68560 23.59149 23.67691 23.53505 23.56674
#>  [48,] 26.39432 26.38744 25.14383 27.04689 26.70654 25.93712
#>  [49,] 23.81685 23.57946 23.61392 23.55862 23.44594 23.66618
#>  [50,] 21.76300 22.04634 22.09226 22.42299 22.52900 22.35539
#>  [51,] 31.83913 31.83195 31.89537 31.64074 31.64583 31.68345
#>  [52,] 31.19671 31.08480 31.26153 31.64551 31.65057 31.68774
#>  [53,] 22.92982 22.89145 22.94163 22.97733 23.32237 22.79634
#>  [54,] 23.11475 22.97586 23.01556 23.64329 23.11671 23.14863
#>  [55,] 23.55508 22.69379      NaN 22.64834 22.84059 22.70440
#>  [56,] 23.17910 23.01995 23.20964 23.08814 23.09396 23.16723
#>  [57,] 22.33415 22.77133 23.89125 23.61941 22.56656 23.33347
#>  [58,]      NaN      NaN      NaN 22.24925 22.92276      NaN
#>  [59,] 23.30953 22.41460      NaN 23.34570 22.88506      NaN
#>  [60,] 22.98155 22.83123 23.10402 23.10018 22.90452 23.07612
#>  [61,] 25.63533 25.63056 25.57016 25.46927 25.71837 26.00041
#>  [62,]      NaN      NaN      NaN      NaN 21.50688 21.79713
#>  [63,] 24.78816 24.75364 24.89947 24.68587 24.76228 24.73593
#>  [64,] 23.36946 22.41970 23.36973 24.29740 24.20527 24.42865
#>  [65,] 24.22724 24.48470 23.99575 24.65502 25.11474 24.91401
#>  [66,]      NaN      NaN 26.66302      NaN      NaN      NaN
#>  [67,] 29.33685 29.23696 29.18880 28.68918 28.35326 28.43538
#>  [68,] 22.61171      NaN      NaN      NaN 22.13082 22.55929
#>  [69,] 21.38221 20.83353 22.31399 21.28416 20.66700 21.40567
#>  [70,] 23.05629 25.77733 24.28101 22.65093      NaN 23.16750
#>  [71,]      NaN      NaN      NaN 24.90495 24.88045 24.87853
#>  [72,] 29.15930 28.97954 29.17268 28.97845 28.94010 28.99323
#>  [73,] 24.81538 24.81108      NaN 25.16273 24.99593 24.48696
#>  [74,]      NaN 20.56039 23.93894      NaN      NaN      NaN
#>  [75,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [76,] 25.64078 25.98383 25.84943 26.03792 26.09563 25.89099
#>  [77,] 23.93625 24.07071 24.16645 23.79412 23.67034 24.14048
#>  [78,]      NaN      NaN      NaN 25.45951      NaN 25.69936
#>  [79,] 23.61717      NaN 23.39218 23.93457 23.81419 24.43993
#>  [80,] 25.04411 24.42284 24.80887 24.75051 26.46516 25.06218
#>  [81,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [82,] 21.01760 18.83102 21.71854 20.03308      NaN      NaN
#>  [83,]      NaN      NaN 23.78137 26.03762 25.82406 25.98372
#>  [84,] 23.01139 23.44380 23.74714 23.46501      NaN 23.98538
#>  [85,] 24.58474      NaN 24.35232 24.71382 24.60761 24.39447
#>  [86,] 25.98407 25.54371 25.70604 25.70588 25.54557 25.47988
#>  [87,] 24.14258      NaN      NaN 23.00520 22.86282 23.29824
#>  [88,] 24.22825 25.19838 25.21892 25.56023 24.54781 24.68239
#>  [89,] 25.25494 25.16158 25.25868 24.48563 24.33572 24.42122
#>  [90,] 23.96570 23.95766 23.53848 23.77485 23.70567 24.09611
#>  [91,] 22.22007 22.55314 22.46971 22.29908 22.71360 22.55880
#>  [92,] 27.07337 26.37496      NaN 26.49095 26.04415 25.82692
#>  [93,]      NaN 22.17411      NaN 22.22955      NaN      NaN
#>  [94,] 25.66668 25.56990 25.62751 25.31979 25.42172 25.37355
#>  [95,] 24.92823 24.26657 25.28207 24.37643 24.60637 24.83615
#>  [96,] 24.91315 24.96886 23.81752 23.81868 23.89989 24.94887
#>  [97,] 24.67023      NaN      NaN 25.22037      NaN 24.77369
#>  [98,] 25.60184 25.44255 25.66162 25.56721 25.35164 26.23320
#>  [99,] 27.72468 27.87121 27.91234 28.12376 27.78094 27.76945
#> [100,] 24.36466 24.49698 24.13406 24.77259 24.48370 24.51431
#> [101,] 20.94149 21.22529 20.85924 21.33642      NaN 21.01475
#> [102,] 25.21119 25.22185 25.26789 25.07521 25.01950 25.06015
#> [103,] 29.26907 29.27143 29.26901 29.34994 29.28869 29.27928
#> [104,] 26.82808 26.46610 27.14287 26.56369 26.43255 26.71463
#> [105,] 27.82771 27.63631 28.25902 27.82343 27.62708 27.93925
#> [106,] 28.64127 28.56216 28.79636 28.58788 28.62931 28.62671
#> [107,] 21.98301 22.13031 20.68526 22.06375 21.79487      NaN
#> [108,]      NaN      NaN      NaN 25.35988 25.25249 25.40720
#> [109,] 24.07423 23.87624 23.38390 23.92325 24.14950 23.85633
#> [110,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [111,]      NaN      NaN 24.80233 24.29614      NaN      NaN
#> [112,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [113,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [114,]      NaN      NaN 23.73345 25.15213 25.10274 23.95409
#> [115,] 24.02776 24.05238 24.04878 24.18867 24.20280 24.30845
#> [116,] 29.79623 29.63568 30.00437 29.71221 29.65475 29.77156
#> [117,] 26.04481 25.96420 26.00243 25.81565 25.96054 25.99252
#> [118,] 23.71691 23.44329 23.63898 23.87802 23.56941 23.80075
#> [119,] 24.05557 24.03347 23.72619 23.76850 23.73665 23.58520
#> [120,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [121,]      NaN      NaN      NaN 24.05292 24.98229 24.47673
#> [122,] 26.61470 26.58663 26.54628 26.57099 26.33451 26.58463
#> [123,] 21.85943 21.77406 21.98676 21.76036 22.29701 22.02759
#> [124,] 28.99110 28.41748 25.98748 28.74679 28.15316 27.27906
#> [125,] 24.51301 24.49550 24.84708 24.63856 24.52152 24.53274
#> [126,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [127,] 24.01683 24.27467 23.99647 24.83014 24.58450 24.91763
#> [128,] 24.04602 24.37206 24.23593 23.59537 23.46438 23.39441
#> [129,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [130,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [131,] 26.69998 26.57738 26.45916 27.16500 27.06394 27.07286
#> [132,] 23.40938      NaN 23.38298 22.30619 23.14659 22.88993
#> [133,]      NaN      NaN 21.15956      NaN      NaN 18.37581
#> [134,] 23.39414 23.20658 23.75328 23.96638 23.05552 23.66021
#> [135,]      NaN      NaN      NaN 22.19761      NaN      NaN
#> [136,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [137,] 22.03708 21.99762 22.28475 22.79677 22.46096 22.53284
#> [138,] 25.36873 25.34841 25.81490 24.63355 24.71288 24.47997
#> [139,] 30.15774 29.97856 30.04934 29.96072 29.84721 29.86616
#> [140,] 25.12589 24.87329 24.78821 25.04802 24.68094 24.81426
#> [141,] 23.47988 22.56193 23.11517 23.10795      NaN 23.77716
#> [142,] 23.88362 23.27739 23.76679 24.13140 23.69434 24.05432
#> [143,] 25.56639 25.76744 25.63156 25.58577 25.63633 25.51078
#> [144,]      NaN 24.75527 23.55039      NaN      NaN      NaN
#> [145,]      NaN      NaN      NaN 24.02724 22.92586      NaN
#> [146,] 26.91593 27.01903 26.83545 26.43212 26.33994 26.24764
#> [147,]      NaN      NaN 22.56302 23.32649 21.88651 22.88889
#> [148,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [149,] 23.79769 23.35676 22.26236 23.66557 22.68895 22.86055
#> [150,] 22.79005      NaN      NaN      NaN      NaN 21.81554
#> [151,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [152,] 23.99895 23.83759 24.12052 24.13782 24.25083 24.09925
#> [153,] 25.75580 25.93933 25.93705 24.64246 23.84929 24.20359
#> [154,] 25.47303 25.41685 25.36904 24.89979 25.47645 25.52155
#> [155,]      NaN      NaN 20.45934      NaN      NaN      NaN
#> [156,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [157,] 22.42611 22.72133 22.80323 23.08463 23.07726 23.27933
#> [158,] 25.54054 25.83918 25.57159 26.41393 26.15888 26.26835
#> [159,] 25.38647 25.10946 25.78980 25.73190 25.67331 25.55277
#> [160,] 21.28901 21.54383 21.81656 21.54727 21.76186 21.61104
#> [161,]      NaN      NaN      NaN 26.22643 26.06122 26.01342
#> [162,] 24.15971 24.06275 24.07118 24.06873 24.14158 24.34434
#> [163,] 27.86526 27.84834 27.99518 27.79017 27.79916 27.82037
#> [164,] 26.49631 26.53921 26.50511 26.53671 26.54206 26.63686
#> [165,]      NaN 22.89770 23.03922 19.83501 22.47417 23.42085
#> [166,] 25.34279      NaN      NaN      NaN      NaN 26.01704
#>  [ getOption("max.print") est atteint -- 504 lignes omises ]