Method to aggregate peptides into proteins with the iterative approach with use of parallelism.

Method to aggregate peptides into proteins with the iterative approach with use of parallelism.

aggIterParallel(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)

Arguments

qPepData
A matrix of intensities of peptides.

X
An adjacency matrix in which lines and columns correspond respectively to peptides and proteins.

conditions
xxxx

init.method
The method used to initialize the iterative algotirhm. Default is 'Sum'.

method
The method used for xxx. Default value is 'Mean'.

n
xxxx

Value

A matrix of intensities of proteins

Author

Samuel Wieczorek

Examples

library(doParallel)

#> Warning: package ‘doParallel’ was built under R version 4.0.2

#> Loading required package: foreach

#> Warning: package ‘foreach’ was built under R version 4.0.2

#> Loading required package: iterators

#> Warning: package ‘iterators’ was built under R version 4.0.2

library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]], 'all'))
conditions <- SummarizedExperiment::colData(obj)$Condition
aggIterParallel(assay(obj,2), X, conditions)

#>      Intensity_C_R1 Intensity_C_R2 Intensity_C_R3 Intensity_D_R1 Intensity_D_R2
#> 1005       24.28821       24.48162       24.32201       24.10576       24.28980
#> 1008       24.23783       23.38166            NaN            NaN            NaN
#> 1015            NaN            NaN            NaN            NaN            NaN
#> 1017       22.50959       22.67980       22.93771       22.55190       22.71791
#> 102        26.09462       26.21429       26.16749       26.09004       26.01065
#> 1024       24.17805       24.10837       23.74755       24.13438       24.88646
#> 103        21.92136       21.55936            NaN       22.78337       22.34627
#> 1033       23.19862       22.03234       23.18264       22.87017       22.59332
#> 1035       27.73982       27.63242       27.89151       27.59328       27.68826
#> 1037       22.01828       22.31921       22.13877       22.46914       22.61715
#> 104             NaN            NaN       21.70092            NaN            NaN
#> 1042       22.62053       23.07604       23.03971       23.12210       22.78965
#> 1046       26.04675       26.16701       26.16127       26.09831       26.08242
#> 105        30.34763       30.31490       30.43516       30.38733       30.38177
#> 1052       23.12366       23.25507       23.25249       23.32313       23.07801
#> 1055            NaN            NaN            NaN            NaN            NaN
#> 1058       23.91031       23.95136       23.77967       24.12305       23.96455
#> 106        29.87954       29.89272       30.03522       30.40243       30.29909
#> 1060       21.31163       21.27509       22.31722       20.93049       21.14582
#> 1063            NaN            NaN            NaN            NaN            NaN
#> 1064       23.73469       23.78726       23.98555            NaN            NaN
#> 107             NaN            NaN            NaN            NaN       22.17283
#> 1073       23.82978       24.91452            NaN            NaN       24.56651
#> 1074       25.62327       25.72882       25.62374       25.57467       25.59867
#> 1077            NaN            NaN            NaN            NaN            NaN
#> 1079       26.03183       26.00093       26.12761       26.00755       25.96299
#> 1080            NaN            NaN            NaN            NaN            NaN
#> 1082       24.66239       24.58734       24.82125       24.67311       24.65929
#> 1086            NaN            NaN            NaN       23.50252       23.54687
#> 1088       24.21076       24.17064       25.04714       24.76779       26.41700
#> 1092            NaN            NaN            NaN            NaN            NaN
#> 1094       23.66839       23.77806       23.64240       24.33470       24.31347
#> 1098       24.93463       24.87037       24.71618       24.91884       24.78905
#> 1099       24.09233       23.91424       24.00273       23.92488       24.03607
#> 1100       23.12307       22.98801       22.78391       22.30048       23.16987
#> 1101       24.76855       24.84460       25.07997       24.80371       24.84575
#> 1102            NaN            NaN            NaN            NaN            NaN
#> 1107       25.29399       25.27184       25.40099       25.43472       25.36470
#> 1109       25.91166       25.37806       26.37814       25.56976       25.30946
#> 111        25.52925       24.96531       25.45001       24.62304       23.68222
#> 1112       25.44038       24.82527       24.95916       25.05887            NaN
#> 1114       28.38187       28.62697       28.24436       28.71954       28.74114
#> 1120       24.30904       24.16507       24.14857       24.55462       24.05606
#> 1121       29.00603       28.83993       29.03708            NaN            NaN
#> 1127       22.74411       23.07084       23.43288       23.80863       23.67347
#> 1129       22.10856       22.91929       22.77429            NaN       23.45237
#> 113        23.50736       23.68560       23.59149       23.67691       23.53505
#> 1134       26.39432       26.38744       25.14383       27.04689       26.70654
#> 1136       23.81685       23.57946       23.61392       23.55862       23.44594
#> 1142       21.76300       22.04634       22.09226       22.42299       22.52900
#> 115        31.83913       31.83195       31.89537       31.64074       31.64583
#> 116        31.19671       31.08480       31.26153       31.64551       31.65057
#> 1161       22.92982       22.89145       22.94163       22.97733       23.32237
#> 1163       23.11475       22.97586       23.01556       23.64329       23.11671
#> 1167       23.55508       22.69379            NaN       22.64834       22.84059
#> 1168       23.17910       23.01995       23.20964       23.08814       23.09396
#> 1169       22.33415       22.77133       23.89125       23.61941       22.56656
#> 1171            NaN            NaN            NaN       22.24925       22.92276
#> 1172       23.30953       22.41460            NaN       23.34570       22.88506
#> 118        22.98155       22.83123       23.10402       23.10018       22.90452
#> 1185       25.63533       25.63056       25.57016       25.46927       25.71837
#> 1187            NaN            NaN            NaN            NaN       21.50688
#> 1189       24.78816       24.75364       24.89947       24.68587       24.76228
#> 1191       23.36946       22.41970       23.36973       24.29740       24.20527
#> 1192       24.22724       24.48470       23.99575       24.65502       25.11474
#> 1198            NaN            NaN       26.66302            NaN            NaN
#> 120        29.33685       29.23696       29.18880       28.68918       28.35326
#> 1202       22.61171            NaN            NaN            NaN       22.13082
#> 1206       21.38221       20.83353       22.31399       21.28416       20.66700
#> 1207       23.05629       25.77733       24.28101       22.65093            NaN
#> 1208            NaN            NaN            NaN       24.90495       24.88045
#> 121        29.15930       28.97954       29.17268       28.97845       28.94010
#> 1210       24.81538       24.81108            NaN       25.16273       24.99593
#> 1212            NaN       20.56039       23.93894            NaN            NaN
#> 1214            NaN            NaN            NaN            NaN            NaN
#> 1216       25.64078       25.98383       25.84943       26.03792       26.09563
#> 1218       23.93625       24.07071       24.16645       23.79412       23.67034
#> 122             NaN            NaN            NaN       25.45951            NaN
#> 1220       23.61717            NaN       23.39218       23.93457       23.81419
#> 1225       25.04411       24.42284       24.80887       24.75051       26.46516
#> 1226            NaN            NaN            NaN            NaN            NaN
#> 1230       21.01760       18.83102       21.71854       20.03308            NaN
#> 1238            NaN            NaN       23.78137       26.03762       25.82406
#> 1239       23.01139       23.44380       23.74714       23.46501            NaN
#> 1245       24.58474            NaN       24.35232       24.71382       24.60761
#> 1246       25.98407       25.54371       25.70604       25.70588       25.54557
#> 1247       24.14258            NaN            NaN       23.00520       22.86282
#> 1251       24.22825       25.19838       25.21892       25.56023       24.54781
#> 1254       25.25494       25.16158       25.25868       24.48563       24.33572
#> 1255       23.96570       23.95766       23.53848       23.77485       23.70567
#> 1257       22.22007       22.55314       22.46971       22.29908       22.71360
#> 1261       27.07337       26.37496            NaN       26.49095       26.04415
#> 1262            NaN       22.17411            NaN       22.22955            NaN
#> 1268       25.66668       25.56990       25.62751       25.31979       25.42172
#> 1277       24.92823       24.26657       25.28207       24.37643       24.60637
#> 1280       24.91315       24.96886       23.81752       23.81868       23.89989
#> 1287       24.67023            NaN            NaN       25.22037            NaN
#> 1288       25.60184       25.44255       25.66162       25.56721       25.35164
#> 129        27.72468       27.87121       27.91234       28.12376       27.78094
#> 1292       24.36466       24.49698       24.13406       24.77259       24.48370
#> 1296       20.94149       21.22529       20.85924       21.33642            NaN
#> 1301       25.21119       25.22185       25.26789       25.07521       25.01950
#> 1304       29.26907       29.27143       29.26901       29.34994       29.28869
#> 1311       26.82808       26.46610       27.14287       26.56369       26.43255
#> 1315       27.82771       27.63631       28.25902       27.82343       27.62708
#> 132        28.64127       28.56216       28.79636       28.58788       28.62931
#> 1320       21.98301       22.13031       20.68526       22.06375       21.79487
#> 1324            NaN            NaN            NaN       25.35988       25.25249
#> 1334       24.07423       23.87624       23.38390       23.92325       24.14950
#> 1337            NaN            NaN            NaN            NaN            NaN
#> 1344            NaN            NaN       24.80233       24.29614            NaN
#> 1348            NaN            NaN            NaN            NaN            NaN
#> 1349            NaN            NaN            NaN            NaN            NaN
#> 1351            NaN            NaN       23.73345       25.15213       25.10274
#> 1356       24.02776       24.05238       24.04878       24.18867       24.20280
#> 1368       29.79623       29.63568       30.00437       29.71221       29.65475
#> 137        26.04481       25.96420       26.00243       25.81565       25.96054
#> 1374       23.71691       23.44329       23.63898       23.87802       23.56941
#> 1384       24.05557       24.03347       23.72619       23.76850       23.73665
#> 1385            NaN            NaN            NaN            NaN            NaN
#> 1392            NaN            NaN            NaN       24.05292       24.98229
#> 1399       26.61470       26.58663       26.54628       26.57099       26.33451
#> 1400       21.85943       21.77406       21.98676       21.76036       22.29701
#> 1405       28.99110       28.41748       25.98748       28.74679       28.15316
#> 1406       24.51301       24.49550       24.84708       24.63856       24.52152
#> 1407            NaN            NaN            NaN            NaN            NaN
#> 1416       24.01683       24.27467       23.99647       24.83014       24.58450
#> 1419       24.04602       24.37206       24.23593       23.59537       23.46438
#> 1421            NaN            NaN            NaN            NaN            NaN
#> 1429            NaN            NaN            NaN            NaN            NaN
#> 143        26.69998       26.57738       26.45916       27.16500       27.06394
#> 1430       23.40938            NaN       23.38298       22.30619       23.14659
#> 1431            NaN            NaN       21.15956            NaN            NaN
#> 1433       23.39414       23.20658       23.75328       23.96638       23.05552
#> 1435            NaN            NaN            NaN       22.19761            NaN
#> 1437            NaN            NaN            NaN            NaN            NaN
#> 1445       22.03708       21.99762       22.28475       22.79677       22.46096
#> 1449       25.36873       25.34841       25.81490       24.63355       24.71288
#> 145        30.15774       29.97856       30.04934       29.96072       29.84721
#> 1452       25.12589       24.87329       24.78821       25.04802       24.68094
#> 1457       23.47988       22.56193       23.11517       23.10795            NaN
#> 146        23.88362       23.27739       23.76679       24.13140       23.69434
#> 1461       25.56639       25.76744       25.63156       25.58577       25.63633
#> 1462            NaN       24.75527       23.55039            NaN            NaN
#> 1463            NaN            NaN            NaN       24.02724       22.92586
#> 1469       26.91593       27.01903       26.83545       26.43212       26.33994
#> 1471            NaN            NaN       22.56302       23.32649       21.88651
#> 1475            NaN            NaN            NaN            NaN            NaN
#> 1478       23.79769       23.35676       22.26236       23.66557       22.68895
#> 1480       22.79005            NaN            NaN            NaN            NaN
#> 149             NaN            NaN            NaN            NaN            NaN
#> 1493       23.99895       23.83759       24.12052       24.13782       24.25083
#> 15         25.75580       25.93933       25.93705       24.64246       23.84929
#> 150        25.47303       25.41685       25.36904       24.89979       25.47645
#> 1506            NaN            NaN       20.45934            NaN            NaN
#> 1510            NaN            NaN            NaN            NaN            NaN
#> 1519       22.42611       22.72133       22.80323       23.08463       23.07726
#> 152        25.54054       25.83918       25.57159       26.41393       26.15888
#> 1523       25.38647       25.10946       25.78980       25.73190       25.67331
#> 1532       21.28901       21.54383       21.81656       21.54727       21.76186
#> 1536            NaN            NaN            NaN       26.22643       26.06122
#> 1548       24.15971       24.06275       24.07118       24.06873       24.14158
#> 155        27.86526       27.84834       27.99518       27.79017       27.79916
#> 1551       26.49631       26.53921       26.50511       26.53671       26.54206
#> 1553            NaN       22.89770       23.03922       19.83501       22.47417
#> 1556       25.34279            NaN            NaN            NaN            NaN
#>      Intensity_D_R3
#> 1005       25.45394
#> 1008            NaN
#> 1015            NaN
#> 1017       22.58516
#> 102        26.09375
#> 1024       24.02547
#> 103             NaN
#> 1033       22.85274
#> 1035       27.73286
#> 1037       22.52920
#> 104             NaN
#> 1042       23.08937
#> 1046       26.03132
#> 105        30.35366
#> 1052       23.35860
#> 1055            NaN
#> 1058       23.86703
#> 106        30.42482
#> 1060       21.36051
#> 1063            NaN
#> 1064            NaN
#> 107             NaN
#> 1073       24.36212
#> 1074       25.66909
#> 1077            NaN
#> 1079       25.92105
#> 1080            NaN
#> 1082       24.70313
#> 1086       23.77957
#> 1088       24.47977
#> 1092            NaN
#> 1094       24.11012
#> 1098       24.80464
#> 1099       24.13312
#> 1100       23.14247
#> 1101       24.95180
#> 1102            NaN
#> 1107       25.15994
#> 1109       25.97235
#> 111        24.56691
#> 1112       25.89809
#> 1114       28.69962
#> 1120       23.84048
#> 1121            NaN
#> 1127       23.63998
#> 1129       22.30265
#> 113        23.56674
#> 1134       25.93712
#> 1136       23.66618
#> 1142       22.35539
#> 115        31.68345
#> 116        31.68774
#> 1161       22.79634
#> 1163       23.14863
#> 1167       22.70440
#> 1168       23.16723
#> 1169       23.33347
#> 1171            NaN
#> 1172            NaN
#> 118        23.07612
#> 1185       26.00041
#> 1187       21.79713
#> 1189       24.73593
#> 1191       24.42865
#> 1192       24.91401
#> 1198            NaN
#> 120        28.43538
#> 1202       22.55929
#> 1206       21.40567
#> 1207       23.16750
#> 1208       24.87853
#> 121        28.99323
#> 1210       24.48696
#> 1212            NaN
#> 1214            NaN
#> 1216       25.89099
#> 1218       24.14048
#> 122        25.69936
#> 1220       24.43993
#> 1225       25.06218
#> 1226            NaN
#> 1230            NaN
#> 1238       25.98372
#> 1239       23.98538
#> 1245       24.39447
#> 1246       25.47988
#> 1247       23.29824
#> 1251       24.68239
#> 1254       24.42122
#> 1255       24.09611
#> 1257       22.55880
#> 1261       25.82692
#> 1262            NaN
#> 1268       25.37355
#> 1277       24.83615
#> 1280       24.94887
#> 1287       24.77369
#> 1288       26.23320
#> 129        27.76945
#> 1292       24.51431
#> 1296       21.01475
#> 1301       25.06015
#> 1304       29.27928
#> 1311       26.71463
#> 1315       27.93925
#> 132        28.62671
#> 1320            NaN
#> 1324       25.40720
#> 1334       23.85633
#> 1337            NaN
#> 1344            NaN
#> 1348            NaN
#> 1349            NaN
#> 1351       23.95409
#> 1356       24.30845
#> 1368       29.77156
#> 137        25.99252
#> 1374       23.80075
#> 1384       23.58520
#> 1385            NaN
#> 1392       24.47673
#> 1399       26.58463
#> 1400       22.02759
#> 1405       27.27906
#> 1406       24.53274
#> 1407            NaN
#> 1416       24.91763
#> 1419       23.39441
#> 1421            NaN
#> 1429            NaN
#> 143        27.07286
#> 1430       22.88993
#> 1431       18.37581
#> 1433       23.66021
#> 1435            NaN
#> 1437            NaN
#> 1445       22.53284
#> 1449       24.47997
#> 145        29.86616
#> 1452       24.81426
#> 1457       23.77716
#> 146        24.05432
#> 1461       25.51078
#> 1462            NaN
#> 1463            NaN
#> 1469       26.24764
#> 1471       22.88889
#> 1475            NaN
#> 1478       22.86055
#> 1480       21.81554
#> 149             NaN
#> 1493       24.09925
#> 15         24.20359
#> 150        25.52155
#> 1506            NaN
#> 1510            NaN
#> 1519       23.27933
#> 152        26.26835
#> 1523       25.55277
#> 1532       21.61104
#> 1536       26.01342
#> 1548       24.34434
#> 155        27.82037
#> 1551       26.63686
#> 1553       23.42085
#> 1556       26.01704
#>  [ getOption("max.print") est atteint -- 504 lignes omises ]