Compute the intensity of proteins with the sum of the intensities of their peptides.

This function computes the intensity of proteins based on the sum of the intensities of their peptides.

aggSum(qPepData, X)

Arguments

qPepData
A matrix of intensities of peptides.

X
An adjacency matrix in which lines and columns correspond respectively to peptides and proteins.

Value

A matrix of intensities of proteins.

Author

Alexia Dorffer, Samuel Wieczorek

Examples

library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]], 'all'))
aggSum(assay(obj[[2]]), X)

#>      Intensity_C_R1 Intensity_C_R2 Intensity_C_R3 Intensity_D_R1 Intensity_D_R2
#> 1005       48.57643       48.96324       48.64402       48.21151       48.57960
#> 1008       24.23783       23.38166        0.00000        0.00000        0.00000
#> 1015        0.00000        0.00000        0.00000        0.00000        0.00000
#> 1017       45.01918       45.35960       45.87542       45.14284       45.00940
#> 102        52.04204       52.19403       52.12040       52.18007       52.02131
#> 1024       24.17805       24.10837       23.74755       24.13438       24.88646
#> 103        21.92136       21.55936        0.00000       22.78337       22.34627
#> 1033       23.19862       22.03234       23.18264       22.87017       22.59332
#> 1035       27.73982       27.63242       27.89151       27.59328       27.68826
#> 1037       22.01828       22.31921       22.13877       22.46914       22.61715
#> 104         0.00000        0.00000       21.70092        0.00000        0.00000
#> 1042       22.62053       23.07604       23.03971       23.12210       22.78965
#> 1046       78.14026       78.50102       78.48382       77.53489       77.27189
#> 105        91.04289       90.94470       91.30547       91.16200       91.14531
#> 1052       46.24733       46.51013       46.50499       46.64625       46.15602
#> 1055        0.00000        0.00000        0.00000        0.00000        0.00000
#> 1058       23.91031       23.95136       23.77967       24.12305       23.96455
#> 106       152.77696      152.74681      153.53367      152.85804      152.78805
#> 1060       21.31163       21.27509       22.31722       20.93049       21.14582
#> 1063       24.57439       24.42458       24.52850       24.03322       23.84048
#> 1064       23.73469       23.78726       23.98555       23.79937       23.56860
#> 107         0.00000        0.00000        0.00000        0.00000       22.17283
#> 1073       23.82978       24.91452        0.00000        0.00000       24.56651
#> 1074       25.62327       25.72882       25.62374       25.57467       25.59867
#> 1077        0.00000       24.28447       24.41796        0.00000       24.00736
#> 1079       26.03183       26.00093       26.12761       26.00755       25.96299
#> 1080       23.74591       23.51605       23.61538        0.00000        0.00000
#> 1082       74.12311       73.59345       74.36368       73.79180       73.82534
#> 1086       23.25767       23.68432       23.90077       23.50252       23.54687
#> 1088       72.63229       72.51193       50.09429       74.30338       26.41700
#> 1092        0.00000        0.00000        0.00000        0.00000       22.96513
#> 1094       23.66839       23.77806       23.64240       24.33470       24.31347
#> 1098       24.93463       24.87037       24.71618       24.91884       24.78905
#> 1099       24.09233       23.91424       24.00273       23.92488       24.03607
#> 1100       23.12307       22.98801       22.78391       22.30048       23.16987
#> 1101       24.76855       24.84460       25.07997       24.80371       24.84575
#> 1102        0.00000       23.34611       23.22985       22.93032       23.28517
#> 1107      128.70213      128.75935      129.07444      129.31194      128.90350
#> 1109       48.19749       48.59404       49.58973       25.56976       25.30946
#> 111       102.27394      100.53806      101.82505      102.43729      100.31233
#> 1112       25.44038       24.82527       24.95916       25.05887        0.00000
#> 1114       28.38187       28.62697       28.24436       28.71954       28.74114
#> 1120       24.30904       24.16507       24.14857       24.55462       24.05606
#> 1121       29.00603       28.83993       29.03708       28.87150       28.83957
#> 1127       22.74411       23.07084       23.43288       23.80863       23.67347
#> 1129       46.89018       47.40723       47.24508       24.31195       48.14208
#> 113        47.69813       47.66270       47.59018       47.76196       47.47453
#> 1134       52.78864       52.77488       50.28766       54.09377       53.41309
#> 1136       23.81685       44.59230       47.03400       47.11724       46.89188
#> 1142       21.76300       22.04634       22.09226       22.42299       22.52900
#> 115        63.36531       63.29992       63.48197       63.28619       63.29634
#> 116        63.36531       63.29992       63.48197       63.28619       63.29634
#> 1161       22.92982       22.89145       22.94163       22.97733       23.32237
#> 1163       46.22950       22.97586       23.01556       47.28658       23.11671
#> 1167       23.55508       22.69379        0.00000       22.64834       22.84059
#> 1168       23.17910       23.01995       23.20964       23.08814       23.09396
#> 1169       22.33415       22.77133       23.89125       23.61941       22.56656
#> 1171        0.00000        0.00000        0.00000       22.24925       22.92276
#> 1172       23.30953       22.41460        0.00000       23.34570       22.88506
#> 118        22.98155       22.83123       23.10402       23.10018       22.90452
#> 1185       25.63533       25.63056       25.57016       25.46927       25.71837
#> 1187        0.00000       21.69799        0.00000        0.00000       21.50688
#> 1189       24.78816       24.75364       24.89947       24.68587       24.76228
#> 1191       49.24425       47.93352       46.90537       48.59480       48.41055
#> 1192       48.45448       48.96939       47.99150       48.72475       25.11474
#> 1198        0.00000        0.00000       26.66302        0.00000        0.00000
#> 120       114.15640      113.54423      112.90300      114.75673      113.41304
#> 1202       22.61171        0.00000        0.00000        0.00000       22.13082
#> 1206       21.38221       20.83353       22.31399       21.28416       20.66700
#> 1207       46.11257       25.77733       48.56203       45.30185        0.00000
#> 1208       25.00350       25.05598       25.18446       24.90495       24.88045
#> 121        58.31860       57.95909       58.34536       57.95690       57.88021
#> 1210       24.81538       24.81108        0.00000       25.16273       24.99593
#> 1212        0.00000       20.56039       23.93894       23.40276        0.00000
#> 1214       24.59628       24.98107       24.52993       24.50246       24.69551
#> 1216       25.64078       25.98383       25.84943       26.03792       26.09563
#> 1218       23.93625       24.07071       24.16645       23.79412       23.67034
#> 122        50.38294       50.11049       25.57793       50.74494       25.33450
#> 1220       47.99824       24.24133       47.64575       47.86914       47.62837
#> 1225       50.08823       48.84567       49.61774       49.50103       26.46516
#> 1226        0.00000        0.00000        0.00000        0.00000        0.00000
#> 1230       44.34780       41.79082       44.58298       43.38487       23.34516
#> 1238       25.81311       25.82369       49.85743       26.03762       25.82406
#> 1239       23.01139       23.44380       23.74714       23.46501        0.00000
#> 1245       24.58474        0.00000       24.35232       24.71382       24.60761
#> 1246       25.98407       25.54371       25.70604       25.70588       25.54557
#> 1247       46.90763       23.16065       23.43403       23.00520       22.86282
#> 1251       24.22825       50.39677       50.43783       51.12047       24.54781
#> 1254       49.42774       49.25777       49.28093       49.27074       49.16807
#> 1255       23.96570       23.95766       23.53848       23.77485       23.70567
#> 1257       22.22007       22.55314       22.46971       22.29908       22.71360
#> 1261       52.73310       52.28790       25.75047       52.98190       52.08830
#> 1262        0.00000       22.17411        0.00000       22.22955        0.00000
#> 1268       25.66668       25.56990       25.62751       25.31979       25.42172
#> 1277       24.92823       24.26657       25.28207       24.37643       24.60637
#> 1280       24.91315       24.96886       47.63504       47.63736       47.79978
#> 1287       24.67023        0.00000        0.00000       25.22037        0.00000
#> 1288       76.80553       76.32764       76.98486       76.70162       76.05493
#> 129        27.72468       27.87121       27.91234       28.12376       27.78094
#> 1292       24.36466       24.49698       24.13406       24.77259       24.48370
#> 1296       20.94149       21.22529       20.85924       21.33642        0.00000
#> 1301       25.21119       25.22185       25.26789       25.07521       25.01950
#> 1304       58.54831       58.55391       58.54865       58.69497       58.57267
#> 1311       54.84015       54.65562       54.79253       54.72691       54.48134
#> 1315       83.45325       82.90756       84.73261       83.49875       82.90788
#> 132       110.49792      109.91800      135.77835      134.71255      135.05194
#> 1320       21.98301       22.13031       20.68526       22.06375       21.79487
#> 1324       25.61243       25.51048       25.66847       25.35988       25.25249
#> 1334       48.00471       47.42886       47.58562       47.79911       47.99237
#> 1337       22.14774        0.00000        0.00000       22.34386        0.00000
#> 1344        0.00000        0.00000       24.80233       24.29614        0.00000
#> 1348       22.43489       22.32877       22.22111       22.39365       22.29953
#> 1349        0.00000       23.20189       23.72078       22.65672       23.26914
#> 1351       25.31060       24.99904       48.97899       25.15213       25.10274
#> 1356       48.05552       48.10476       48.09755       48.37734       48.40559
#> 1368       29.79623       29.63568       30.00437       29.71221       29.65475
#> 137        26.04481       25.96420       26.00243       25.81565       25.96054
#> 1374       23.71691       23.44329       23.63898       23.87802       23.56941
#> 1384       24.05557       24.03347       23.72619       23.76850       23.73665
#> 1385       26.36896       26.46941       26.67438       26.44969       26.52453
#> 1392       24.73939       25.42191       24.70356       24.05292       24.98229
#> 1399       26.61470       26.58663       26.54628       26.57099       26.33451
#> 1400       21.85943       21.77406       21.98676       21.76036       22.29701
#> 1405       28.99110       28.41748       25.98748       28.74679       28.15316
#> 1406       73.53903       73.48651       74.54123       73.91569       73.56457
#> 1407       21.67990       21.36538        0.00000        0.00000        0.00000
#> 1416       72.05048       48.54934       71.98942       49.66029       49.16901
#> 1419       24.04602       24.37206       24.23593       23.59537       23.46438
#> 1421        0.00000        0.00000        0.00000        0.00000        0.00000
#> 1429        0.00000        0.00000        0.00000        0.00000        0.00000
#> 143       108.23977      108.00173      107.82539      109.49112      108.27689
#> 1430       23.40938        0.00000       23.38298       22.30619       23.14659
#> 1431        0.00000        0.00000       21.15956        0.00000        0.00000
#> 1433       70.18241       93.79650       71.93522       47.93277       69.16655
#> 1435        0.00000        0.00000        0.00000       22.19761        0.00000
#> 1437       22.12729        0.00000        0.00000        0.00000        0.00000
#> 1445       22.03708       21.99762       22.28475       22.79677       22.46096
#> 1449       49.39234       49.47934       49.42296       49.26710       49.42576
#> 145        30.15774       29.97856       30.04934       29.96072       29.84721
#> 1452       25.12589       24.87329       24.78821       25.04802       24.68094
#> 1457       23.47988       22.56193       23.11517       46.21590        0.00000
#> 146        23.88362       23.27739       23.76679       24.13140       23.69434
#> 1461       25.56639       25.76744       25.63156       25.58577       25.63633
#> 1462        0.00000       24.75527       23.55039       23.66503       24.07235
#> 1463        0.00000        0.00000        0.00000       24.02724       22.92586
#> 1469       79.01047       79.18571       78.88209       79.29637       79.01981
#> 1471        0.00000        0.00000       22.56302       23.32649       21.88651
#> 1475       23.20829       23.02800       23.36065       23.17676       22.85302
#> 1478       23.79769       23.35676       22.26236       23.66557       22.68895
#> 1480       22.79005        0.00000        0.00000        0.00000        0.00000
#> 149        25.92019       25.57661       26.17287       25.80058       25.69064
#> 1493       23.99895       23.83759       24.12052       24.13782       24.25083
#> 15         77.79256       78.22043       78.26002       48.55222       48.49444
#> 150       152.83819      152.50113      177.58328      174.29854      152.85870
#> 1506        0.00000        0.00000       20.45934       21.35396        0.00000
#> 1510       23.27583       23.06064       23.25465       22.67588       23.06615
#> 1519       44.85223       45.44265       45.60647       46.16926       46.15452
#> 152        51.08108       51.67835       51.14317       52.82786       52.31776
#> 1523       50.77294       50.21892       25.78980       50.50951       50.14905
#> 1532       21.28901       21.54383       21.81656       21.54727       21.76186
#> 1536       26.18595       26.05403       26.21969       26.22643       26.06122
#> 1548       48.31941       48.12549       48.14237       48.13745       48.28316
#> 155        27.86526       27.84834       27.99518       27.79017       27.79916
#> 1551       26.49631       26.53921       26.50511       26.53671       26.54206
#> 1553        0.00000       22.89770       23.03922       19.83501       22.47417
#> 1556       25.34279        0.00000        0.00000        0.00000        0.00000
#>      Intensity_D_R3
#> 1005       25.45394
#> 1008        0.00000
#> 1015        0.00000
#> 1017       45.52801
#> 102        52.18750
#> 1024       24.02547
#> 103         0.00000
#> 1033       22.85274
#> 1035       27.73286
#> 1037       22.52920
#> 104         0.00000
#> 1042       23.08937
#> 1046       77.39914
#> 105        91.06098
#> 1052       46.71719
#> 1055        0.00000
#> 1058       23.86703
#> 106       153.17815
#> 1060       21.36051
#> 1063       24.22074
#> 1064       23.82172
#> 107         0.00000
#> 1073       24.36212
#> 1074       25.66909
#> 1077        0.00000
#> 1079       25.92105
#> 1080        0.00000
#> 1082       74.07928
#> 1086       23.77957
#> 1088       73.43931
#> 1092       22.83836
#> 1094       24.11012
#> 1098       24.80464
#> 1099       24.13312
#> 1100       23.14247
#> 1101       24.95180
#> 1102       23.13646
#> 1107      128.55322
#> 1109       49.67189
#> 111       102.39926
#> 1112       25.89809
#> 1114       28.69962
#> 1120       23.84048
#> 1121       28.93340
#> 1127       23.63998
#> 1129       46.82213
#> 113        47.47449
#> 1134       51.87424
#> 1136       47.33236
#> 1142       22.35539
#> 115        63.37114
#> 116        63.37114
#> 1161       22.79634
#> 1163       23.14863
#> 1167       22.70440
#> 1168       23.16723
#> 1169       23.33347
#> 1171        0.00000
#> 1172        0.00000
#> 118        23.07612
#> 1185       26.00041
#> 1187       21.79713
#> 1189       24.73593
#> 1191       48.85729
#> 1192       48.55443
#> 1198        0.00000
#> 120       113.74153
#> 1202       22.55929
#> 1206       21.40567
#> 1207       46.33499
#> 1208       24.87853
#> 121        57.98646
#> 1210       24.48696
#> 1212       23.14355
#> 1214       24.70883
#> 1216       25.89099
#> 1218       24.14048
#> 122        25.69936
#> 1220       24.43993
#> 1225       50.12437
#> 1226        0.00000
#> 1230       23.21267
#> 1238       25.98372
#> 1239       23.98538
#> 1245       24.39447
#> 1246       25.47988
#> 1247       23.29824
#> 1251       24.68239
#> 1254       49.25341
#> 1255       24.09611
#> 1257       22.55880
#> 1261       51.65384
#> 1262        0.00000
#> 1268       25.37355
#> 1277       24.83615
#> 1280       24.94887
#> 1287       24.77369
#> 1288       78.69959
#> 129        27.76945
#> 1292       24.51431
#> 1296       21.01475
#> 1301       25.06015
#> 1304       58.55439
#> 1311       54.60771
#> 1315       83.83535
#> 132       134.87554
#> 1320        0.00000
#> 1324       25.40720
#> 1334       47.94406
#> 1337        0.00000
#> 1344        0.00000
#> 1348       22.42121
#> 1349       22.79098
#> 1351       47.90818
#> 1356       48.61690
#> 1368       29.77156
#> 137        25.99252
#> 1374       23.80075
#> 1384       23.58520
#> 1385       26.40827
#> 1392       24.47673
#> 1399       26.58463
#> 1400       22.02759
#> 1405       27.27906
#> 1406       73.59821
#> 1407        0.00000
#> 1416       49.83526
#> 1419       23.39441
#> 1421        0.00000
#> 1429        0.00000
#> 143       108.46694
#> 1430       22.88993
#> 1431       18.37581
#> 1433       94.64085
#> 1435        0.00000
#> 1437        0.00000
#> 1445       22.53284
#> 1449       48.95993
#> 145        29.86616
#> 1452       24.81426
#> 1457       23.77716
#> 146        24.05432
#> 1461       25.51078
#> 1462        0.00000
#> 1463        0.00000
#> 1469       78.74293
#> 1471       22.88889
#> 1475       22.90121
#> 1478       22.86055
#> 1480       21.81554
#> 149        25.72289
#> 1493       24.09925
#> 15         72.23579
#> 150       153.12929
#> 1506       20.53744
#> 1510       23.05243
#> 1519       46.55867
#> 152        52.53670
#> 1523       50.31104
#> 1532       21.61104
#> 1536       26.01342
#> 1548       48.68867
#> 155        27.82037
#> 1551       26.63686
#> 1553       23.42085
#> 1556       26.01704
#>  [ getOption("max.print") est atteint -- 504 lignes omises ]