Aggreagate quantitative features.

This function takes a matrix of quantitative features `x` and a factor (of length equal to `nrow(x)`) defining subsets, and applies a user-defined function to aggregate each subset into a vector of quantitative values. This function is the same as 'aggregate_by_vector' from the package QFeatures and it is used with DAPAR aggregation functions

User-defined functions must thus return a matrix of dimensions equal to `X`. Examples thereof are

- [DAPAR2:aggSum()] to use the sum of each column (default);

- [DAPAR2:aggMean()] to use the sum of each column;

- [DAPAR2:aggIter()] to use the mean of each column;

- [DAPAR2:aggIterParallel()] same as previous function but use parallelism.

- [DAPAR::aggTopn] to use the sum of each column;

aggregate_with_matAdj(qPepData, X, FUN, ...)

Arguments

qPepData
xxxx.

X
An adjacency matrix.

FUN
A `function` to be applied to the subsets of `x`.

...
Additional arguments passed to `FUN`.

Value

A new `matrix` of dimensions xxxx

Author

Samuel Wieczorek