Creates an object of class QFeatures from an object of class MSnSet

xxxx

convertMSnset2QFeatures(
  obj,
  analysis,
  parentProtId,
  keyId,
  pipelineType = NULL,
  processes = NULL
)

Arguments

obj
xxx.

analysis
xxx

parentProtId
For peptide entities, a string which is the name of a column in rowData. It contains the id of parent proteins and is used to generate adjacency matrix and process to aggregation.

keyId
The indice of the column containing the ID of entities (peptides or proteins)

pipelineType
xxxx

processes
xxxx

Value

An instance of class QFeatures.

Author

Samuel Wieczorek

Examples

library(QFeatures)
library(MSnbase)

#> Le chargement a nécessité le package : mzR

#> Le chargement a nécessité le package : Rcpp

#> Warning: le package 'Rcpp' a été compilé avec la version R 4.0.2

#> Warning: mzR has been built against a different Rcpp version (1.0.4.6)
#> than is installed on your system (1.0.5). This might lead to errors
#> when loading mzR. If you encounter such issues, please send a report,
#> including the output of sessionInfo() to the Bioc support forum at 
#> https://support.bioconductor.org/. For details see also
#> https://github.com/sneumann/mzR/wiki/mzR-Rcpp-compiler-linker-issue.

#> Le chargement a nécessité le package : ProtGenerics

#> 
#> Attachement du package : 'ProtGenerics'

#> The following object is masked from 'package:stats':
#> 
#>     smooth

#> 
#> This is MSnbase version 2.14.2 
#>   Visit https://lgatto.github.io/MSnbase/ to get started.

#> 
#> Attachement du package : 'MSnbase'

#> The following object is masked from 'package:base':
#> 
#>     trimws

utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
parentId <- 'Protein_group_IDs'
keyid <- 'Sequence'
ft <- convertMSnset2QFeatures(obj, 'conv',parentId, keyid )