Similar to the function is.na but focused on the equality with the paramter 'type'.

Similar to the function is.na but focused on the equality with the paramter 'type'.

is.OfType(data, type)

Arguments

data
A data.frame

type
The value to search in the dataframe

Value

A boolean dataframe

Author

Samuel Wieczorek

Examples

library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept
data <- rowData(Exp1_R25_pept[['original']])[,metadata(Exp1_R25_pept)$OriginOfValues]
is.OfType(as.data.frame(data@listData), "MEC")

#>          OriginOfValue_Intensity_C_R1 OriginOfValue_Intensity_C_R2
#>     [1,]                        FALSE                        FALSE
#>     [2,]                        FALSE                        FALSE
#>     [3,]                        FALSE                        FALSE
#>     [4,]                        FALSE                        FALSE
#>     [5,]                        FALSE                        FALSE
#>     [6,]                        FALSE                        FALSE
#>     [7,]                        FALSE                        FALSE
#>     [8,]                        FALSE                        FALSE
#>     [9,]                        FALSE                        FALSE
#>    [10,]                        FALSE                        FALSE
#>    [11,]                        FALSE                        FALSE
#>    [12,]                        FALSE                        FALSE
#>    [13,]                        FALSE                        FALSE
#>    [14,]                        FALSE                        FALSE
#>    [15,]                        FALSE                        FALSE
#>    [16,]                        FALSE                        FALSE
#>    [17,]                        FALSE                        FALSE
#>    [18,]                        FALSE                        FALSE
#>    [19,]                        FALSE                        FALSE
#>    [20,]                        FALSE                        FALSE
#>    [21,]                        FALSE                        FALSE
#>    [22,]                        FALSE                        FALSE
#>    [23,]                         TRUE                         TRUE
#>    [24,]                        FALSE                        FALSE
#>    [25,]                        FALSE                        FALSE
#>    [26,]                         TRUE                         TRUE
#>    [27,]                        FALSE                        FALSE
#>    [28,]                        FALSE                        FALSE
#>    [29,]                        FALSE                        FALSE
#>    [30,]                        FALSE                        FALSE
#>    [31,]                        FALSE                        FALSE
#>    [32,]                        FALSE                        FALSE
#>    [33,]                        FALSE                        FALSE
#>    [34,]                        FALSE                        FALSE
#>    [35,]                        FALSE                        FALSE
#>    [36,]                        FALSE                        FALSE
#>    [37,]                        FALSE                        FALSE
#>    [38,]                        FALSE                        FALSE
#>    [39,]                        FALSE                        FALSE
#>    [40,]                        FALSE                        FALSE
#>    [41,]                        FALSE                        FALSE
#>    [42,]                        FALSE                        FALSE
#>    [43,]                         TRUE                         TRUE
#>    [44,]                        FALSE                        FALSE
#>    [45,]                         TRUE                         TRUE
#>    [46,]                        FALSE                        FALSE
#>    [47,]                        FALSE                        FALSE
#>    [48,]                        FALSE                        FALSE
#>    [49,]                        FALSE                        FALSE
#>    [50,]                        FALSE                        FALSE
#>    [51,]                        FALSE                        FALSE
#>    [52,]                        FALSE                        FALSE
#>    [53,]                        FALSE                        FALSE
#>    [54,]                        FALSE                        FALSE
#>    [55,]                        FALSE                        FALSE
#>    [56,]                        FALSE                        FALSE
#>    [57,]                        FALSE                        FALSE
#>    [58,]                        FALSE                        FALSE
#>    [59,]                        FALSE                        FALSE
#>    [60,]                        FALSE                        FALSE
#>    [61,]                        FALSE                        FALSE
#>    [62,]                        FALSE                        FALSE
#>    [63,]                         TRUE                         TRUE
#>    [64,]                        FALSE                        FALSE
#>    [65,]                        FALSE                        FALSE
#>    [66,]                        FALSE                        FALSE
#>    [67,]                        FALSE                        FALSE
#>    [68,]                        FALSE                        FALSE
#>    [69,]                        FALSE                        FALSE
#>    [70,]                        FALSE                        FALSE
#>    [71,]                        FALSE                        FALSE
#>    [72,]                        FALSE                        FALSE
#>    [73,]                        FALSE                        FALSE
#>    [74,]                        FALSE                        FALSE
#>    [75,]                        FALSE                        FALSE
#>    [76,]                        FALSE                        FALSE
#>    [77,]                        FALSE                        FALSE
#>    [78,]                        FALSE                        FALSE
#>    [79,]                        FALSE                        FALSE
#>    [80,]                        FALSE                        FALSE
#>    [81,]                        FALSE                        FALSE
#>    [82,]                        FALSE                        FALSE
#>    [83,]                        FALSE                        FALSE
#>    [84,]                        FALSE                        FALSE
#>    [85,]                        FALSE                        FALSE
#>    [86,]                        FALSE                        FALSE
#>    [87,]                        FALSE                        FALSE
#>    [88,]                        FALSE                        FALSE
#>    [89,]                        FALSE                        FALSE
#>    [90,]                        FALSE                        FALSE
#>    [91,]                        FALSE                        FALSE
#>    [92,]                        FALSE                        FALSE
#>    [93,]                         TRUE                         TRUE
#>    [94,]                        FALSE                        FALSE
#>    [95,]                        FALSE                        FALSE
#>    [96,]                        FALSE                        FALSE
#>    [97,]                        FALSE                        FALSE
#>    [98,]                        FALSE                        FALSE
#>    [99,]                        FALSE                        FALSE
#>   [100,]                        FALSE                        FALSE
#>   [101,]                         TRUE                         TRUE
#>   [102,]                        FALSE                        FALSE
#>   [103,]                        FALSE                        FALSE
#>   [104,]                        FALSE                        FALSE
#>   [105,]                        FALSE                        FALSE
#>   [106,]                        FALSE                        FALSE
#>   [107,]                        FALSE                        FALSE
#>   [108,]                        FALSE                        FALSE
#>   [109,]                        FALSE                        FALSE
#>   [110,]                        FALSE                        FALSE
#>   [111,]                        FALSE                        FALSE
#>   [112,]                        FALSE                        FALSE
#>   [113,]                        FALSE                        FALSE
#>   [114,]                        FALSE                        FALSE
#>   [115,]                        FALSE                        FALSE
#>   [116,]                        FALSE                        FALSE
#>   [117,]                        FALSE                        FALSE
#>   [118,]                        FALSE                        FALSE
#>   [119,]                        FALSE                        FALSE
#>   [120,]                        FALSE                        FALSE
#>   [121,]                        FALSE                        FALSE
#>   [122,]                        FALSE                        FALSE
#>   [123,]                        FALSE                        FALSE
#>   [124,]                        FALSE                        FALSE
#>   [125,]                        FALSE                        FALSE
#>   [126,]                        FALSE                        FALSE
#>   [127,]                        FALSE                        FALSE
#>   [128,]                        FALSE                        FALSE
#>   [129,]                        FALSE                        FALSE
#>   [130,]                        FALSE                        FALSE
#>   [131,]                        FALSE                        FALSE
#>   [132,]                        FALSE                        FALSE
#>   [133,]                        FALSE                        FALSE
#>   [134,]                        FALSE                        FALSE
#>   [135,]                        FALSE                        FALSE
#>   [136,]                        FALSE                        FALSE
#>   [137,]                        FALSE                        FALSE
#>   [138,]                        FALSE                        FALSE
#>   [139,]                        FALSE                        FALSE
#>   [140,]                        FALSE                        FALSE
#>   [141,]                        FALSE                        FALSE
#>   [142,]                        FALSE                        FALSE
#>   [143,]                        FALSE                        FALSE
#>   [144,]                        FALSE                        FALSE
#>   [145,]                         TRUE                         TRUE
#>   [146,]                        FALSE                        FALSE
#>   [147,]                        FALSE                        FALSE
#>   [148,]                        FALSE                        FALSE
#>   [149,]                        FALSE                        FALSE
#>   [150,]                        FALSE                        FALSE
#>   [151,]                        FALSE                        FALSE
#>   [152,]                        FALSE                        FALSE
#>   [153,]                        FALSE                        FALSE
#>   [154,]                        FALSE                        FALSE
#>   [155,]                        FALSE                        FALSE
#>   [156,]                        FALSE                        FALSE
#>   [157,]                        FALSE                        FALSE
#>   [158,]                        FALSE                        FALSE
#>   [159,]                        FALSE                        FALSE
#>   [160,]                        FALSE                        FALSE
#>   [161,]                        FALSE                        FALSE
#>   [162,]                        FALSE                        FALSE
#>   [163,]                        FALSE                        FALSE
#>   [164,]                        FALSE                        FALSE
#>   [165,]                        FALSE                        FALSE
#>   [166,]                        FALSE                        FALSE
#>          OriginOfValue_Intensity_C_R3 OriginOfValue_Intensity_D_R1
#>     [1,]                        FALSE                        FALSE
#>     [2,]                        FALSE                        FALSE
#>     [3,]                        FALSE                        FALSE
#>     [4,]                        FALSE                         TRUE
#>     [5,]                        FALSE                        FALSE
#>     [6,]                        FALSE                        FALSE
#>     [7,]                        FALSE                        FALSE
#>     [8,]                        FALSE                        FALSE
#>     [9,]                        FALSE                        FALSE
#>    [10,]                        FALSE                        FALSE
#>    [11,]                        FALSE                        FALSE
#>    [12,]                        FALSE                        FALSE
#>    [13,]                        FALSE                        FALSE
#>    [14,]                        FALSE                         TRUE
#>    [15,]                        FALSE                        FALSE
#>    [16,]                        FALSE                        FALSE
#>    [17,]                        FALSE                        FALSE
#>    [18,]                        FALSE                        FALSE
#>    [19,]                        FALSE                        FALSE
#>    [20,]                        FALSE                        FALSE
#>    [21,]                        FALSE                        FALSE
#>    [22,]                        FALSE                        FALSE
#>    [23,]                         TRUE                         TRUE
#>    [24,]                        FALSE                        FALSE
#>    [25,]                        FALSE                        FALSE
#>    [26,]                         TRUE                         TRUE
#>    [27,]                        FALSE                        FALSE
#>    [28,]                        FALSE                        FALSE
#>    [29,]                        FALSE                        FALSE
#>    [30,]                        FALSE                         TRUE
#>    [31,]                        FALSE                        FALSE
#>    [32,]                        FALSE                        FALSE
#>    [33,]                        FALSE                        FALSE
#>    [34,]                        FALSE                        FALSE
#>    [35,]                        FALSE                        FALSE
#>    [36,]                        FALSE                        FALSE
#>    [37,]                        FALSE                        FALSE
#>    [38,]                        FALSE                        FALSE
#>    [39,]                        FALSE                        FALSE
#>    [40,]                        FALSE                        FALSE
#>    [41,]                        FALSE                        FALSE
#>    [42,]                        FALSE                        FALSE
#>    [43,]                         TRUE                         TRUE
#>    [44,]                        FALSE                        FALSE
#>    [45,]                         TRUE                        FALSE
#>    [46,]                        FALSE                         TRUE
#>    [47,]                        FALSE                        FALSE
#>    [48,]                        FALSE                        FALSE
#>    [49,]                        FALSE                        FALSE
#>    [50,]                        FALSE                        FALSE
#>    [51,]                        FALSE                         TRUE
#>    [52,]                        FALSE                        FALSE
#>    [53,]                        FALSE                        FALSE
#>    [54,]                        FALSE                        FALSE
#>    [55,]                        FALSE                        FALSE
#>    [56,]                        FALSE                        FALSE
#>    [57,]                        FALSE                        FALSE
#>    [58,]                        FALSE                        FALSE
#>    [59,]                        FALSE                        FALSE
#>    [60,]                        FALSE                        FALSE
#>    [61,]                        FALSE                        FALSE
#>    [62,]                        FALSE                        FALSE
#>    [63,]                         TRUE                        FALSE
#>    [64,]                        FALSE                        FALSE
#>    [65,]                        FALSE                        FALSE
#>    [66,]                        FALSE                        FALSE
#>    [67,]                        FALSE                        FALSE
#>    [68,]                        FALSE                        FALSE
#>    [69,]                        FALSE                        FALSE
#>    [70,]                        FALSE                        FALSE
#>    [71,]                        FALSE                        FALSE
#>    [72,]                        FALSE                        FALSE
#>    [73,]                        FALSE                        FALSE
#>    [74,]                        FALSE                        FALSE
#>    [75,]                        FALSE                        FALSE
#>    [76,]                        FALSE                        FALSE
#>    [77,]                        FALSE                        FALSE
#>    [78,]                        FALSE                        FALSE
#>    [79,]                        FALSE                        FALSE
#>    [80,]                        FALSE                        FALSE
#>    [81,]                        FALSE                        FALSE
#>    [82,]                        FALSE                        FALSE
#>    [83,]                        FALSE                        FALSE
#>    [84,]                        FALSE                        FALSE
#>    [85,]                        FALSE                        FALSE
#>    [86,]                        FALSE                        FALSE
#>    [87,]                        FALSE                        FALSE
#>    [88,]                        FALSE                        FALSE
#>    [89,]                        FALSE                        FALSE
#>    [90,]                        FALSE                        FALSE
#>    [91,]                        FALSE                        FALSE
#>    [92,]                        FALSE                        FALSE
#>    [93,]                         TRUE                         TRUE
#>    [94,]                        FALSE                        FALSE
#>    [95,]                        FALSE                        FALSE
#>    [96,]                        FALSE                        FALSE
#>    [97,]                        FALSE                        FALSE
#>    [98,]                        FALSE                        FALSE
#>    [99,]                        FALSE                        FALSE
#>   [100,]                        FALSE                        FALSE
#>   [101,]                         TRUE                         TRUE
#>   [102,]                        FALSE                        FALSE
#>   [103,]                        FALSE                        FALSE
#>   [104,]                        FALSE                        FALSE
#>   [105,]                        FALSE                        FALSE
#>   [106,]                        FALSE                         TRUE
#>   [107,]                        FALSE                        FALSE
#>   [108,]                        FALSE                        FALSE
#>   [109,]                        FALSE                         TRUE
#>   [110,]                        FALSE                        FALSE
#>   [111,]                        FALSE                        FALSE
#>   [112,]                        FALSE                        FALSE
#>   [113,]                        FALSE                        FALSE
#>   [114,]                        FALSE                        FALSE
#>   [115,]                        FALSE                        FALSE
#>   [116,]                        FALSE                        FALSE
#>   [117,]                        FALSE                        FALSE
#>   [118,]                        FALSE                        FALSE
#>   [119,]                        FALSE                        FALSE
#>   [120,]                        FALSE                        FALSE
#>   [121,]                        FALSE                        FALSE
#>   [122,]                        FALSE                        FALSE
#>   [123,]                        FALSE                        FALSE
#>   [124,]                        FALSE                        FALSE
#>   [125,]                        FALSE                        FALSE
#>   [126,]                        FALSE                        FALSE
#>   [127,]                        FALSE                        FALSE
#>   [128,]                        FALSE                        FALSE
#>   [129,]                        FALSE                        FALSE
#>   [130,]                        FALSE                        FALSE
#>   [131,]                        FALSE                        FALSE
#>   [132,]                        FALSE                        FALSE
#>   [133,]                        FALSE                        FALSE
#>   [134,]                        FALSE                        FALSE
#>   [135,]                        FALSE                        FALSE
#>   [136,]                        FALSE                        FALSE
#>   [137,]                        FALSE                        FALSE
#>   [138,]                        FALSE                        FALSE
#>   [139,]                        FALSE                        FALSE
#>   [140,]                        FALSE                        FALSE
#>   [141,]                        FALSE                        FALSE
#>   [142,]                        FALSE                        FALSE
#>   [143,]                        FALSE                        FALSE
#>   [144,]                        FALSE                        FALSE
#>   [145,]                         TRUE                         TRUE
#>   [146,]                        FALSE                        FALSE
#>   [147,]                        FALSE                         TRUE
#>   [148,]                        FALSE                        FALSE
#>   [149,]                        FALSE                        FALSE
#>   [150,]                        FALSE                        FALSE
#>   [151,]                        FALSE                        FALSE
#>   [152,]                        FALSE                        FALSE
#>   [153,]                        FALSE                        FALSE
#>   [154,]                        FALSE                        FALSE
#>   [155,]                        FALSE                        FALSE
#>   [156,]                        FALSE                        FALSE
#>   [157,]                        FALSE                        FALSE
#>   [158,]                        FALSE                        FALSE
#>   [159,]                        FALSE                        FALSE
#>   [160,]                        FALSE                        FALSE
#>   [161,]                        FALSE                        FALSE
#>   [162,]                        FALSE                        FALSE
#>   [163,]                        FALSE                        FALSE
#>   [164,]                        FALSE                        FALSE
#>   [165,]                        FALSE                        FALSE
#>   [166,]                        FALSE                        FALSE
#>          OriginOfValue_Intensity_D_R2 OriginOfValue_Intensity_D_R3
#>     [1,]                        FALSE                        FALSE
#>     [2,]                        FALSE                        FALSE
#>     [3,]                        FALSE                        FALSE
#>     [4,]                         TRUE                         TRUE
#>     [5,]                        FALSE                        FALSE
#>     [6,]                        FALSE                        FALSE
#>     [7,]                        FALSE                        FALSE
#>     [8,]                        FALSE                        FALSE
#>     [9,]                        FALSE                        FALSE
#>    [10,]                        FALSE                        FALSE
#>    [11,]                        FALSE                        FALSE
#>    [12,]                        FALSE                        FALSE
#>    [13,]                        FALSE                        FALSE
#>    [14,]                         TRUE                         TRUE
#>    [15,]                        FALSE                        FALSE
#>    [16,]                        FALSE                        FALSE
#>    [17,]                        FALSE                        FALSE
#>    [18,]                        FALSE                        FALSE
#>    [19,]                        FALSE                        FALSE
#>    [20,]                        FALSE                        FALSE
#>    [21,]                        FALSE                        FALSE
#>    [22,]                        FALSE                        FALSE
#>    [23,]                         TRUE                         TRUE
#>    [24,]                        FALSE                        FALSE
#>    [25,]                        FALSE                        FALSE
#>    [26,]                         TRUE                         TRUE
#>    [27,]                        FALSE                        FALSE
#>    [28,]                        FALSE                        FALSE
#>    [29,]                        FALSE                        FALSE
#>    [30,]                         TRUE                         TRUE
#>    [31,]                        FALSE                        FALSE
#>    [32,]                        FALSE                        FALSE
#>    [33,]                        FALSE                        FALSE
#>    [34,]                        FALSE                        FALSE
#>    [35,]                        FALSE                        FALSE
#>    [36,]                        FALSE                        FALSE
#>    [37,]                        FALSE                        FALSE
#>    [38,]                        FALSE                        FALSE
#>    [39,]                        FALSE                        FALSE
#>    [40,]                        FALSE                        FALSE
#>    [41,]                        FALSE                        FALSE
#>    [42,]                        FALSE                        FALSE
#>    [43,]                         TRUE                         TRUE
#>    [44,]                        FALSE                        FALSE
#>    [45,]                        FALSE                        FALSE
#>    [46,]                         TRUE                         TRUE
#>    [47,]                        FALSE                        FALSE
#>    [48,]                        FALSE                        FALSE
#>    [49,]                        FALSE                        FALSE
#>    [50,]                        FALSE                        FALSE
#>    [51,]                         TRUE                         TRUE
#>    [52,]                        FALSE                        FALSE
#>    [53,]                        FALSE                        FALSE
#>    [54,]                        FALSE                        FALSE
#>    [55,]                        FALSE                        FALSE
#>    [56,]                        FALSE                        FALSE
#>    [57,]                        FALSE                        FALSE
#>    [58,]                        FALSE                        FALSE
#>    [59,]                        FALSE                        FALSE
#>    [60,]                        FALSE                        FALSE
#>    [61,]                        FALSE                        FALSE
#>    [62,]                        FALSE                        FALSE
#>    [63,]                        FALSE                        FALSE
#>    [64,]                        FALSE                        FALSE
#>    [65,]                        FALSE                        FALSE
#>    [66,]                        FALSE                        FALSE
#>    [67,]                        FALSE                        FALSE
#>    [68,]                        FALSE                        FALSE
#>    [69,]                        FALSE                        FALSE
#>    [70,]                        FALSE                        FALSE
#>    [71,]                        FALSE                        FALSE
#>    [72,]                        FALSE                        FALSE
#>    [73,]                        FALSE                        FALSE
#>    [74,]                        FALSE                        FALSE
#>    [75,]                        FALSE                        FALSE
#>    [76,]                        FALSE                        FALSE
#>    [77,]                        FALSE                        FALSE
#>    [78,]                        FALSE                        FALSE
#>    [79,]                        FALSE                        FALSE
#>    [80,]                        FALSE                        FALSE
#>    [81,]                        FALSE                        FALSE
#>    [82,]                        FALSE                        FALSE
#>    [83,]                        FALSE                        FALSE
#>    [84,]                        FALSE                        FALSE
#>    [85,]                        FALSE                        FALSE
#>    [86,]                        FALSE                        FALSE
#>    [87,]                        FALSE                        FALSE
#>    [88,]                        FALSE                        FALSE
#>    [89,]                        FALSE                        FALSE
#>    [90,]                        FALSE                        FALSE
#>    [91,]                        FALSE                        FALSE
#>    [92,]                        FALSE                        FALSE
#>    [93,]                         TRUE                         TRUE
#>    [94,]                        FALSE                        FALSE
#>    [95,]                        FALSE                        FALSE
#>    [96,]                        FALSE                        FALSE
#>    [97,]                        FALSE                        FALSE
#>    [98,]                        FALSE                        FALSE
#>    [99,]                        FALSE                        FALSE
#>   [100,]                        FALSE                        FALSE
#>   [101,]                         TRUE                         TRUE
#>   [102,]                        FALSE                        FALSE
#>   [103,]                        FALSE                        FALSE
#>   [104,]                        FALSE                        FALSE
#>   [105,]                        FALSE                        FALSE
#>   [106,]                         TRUE                         TRUE
#>   [107,]                        FALSE                        FALSE
#>   [108,]                        FALSE                        FALSE
#>   [109,]                         TRUE                         TRUE
#>   [110,]                        FALSE                        FALSE
#>   [111,]                        FALSE                        FALSE
#>   [112,]                        FALSE                        FALSE
#>   [113,]                        FALSE                        FALSE
#>   [114,]                        FALSE                        FALSE
#>   [115,]                        FALSE                        FALSE
#>   [116,]                        FALSE                        FALSE
#>   [117,]                        FALSE                        FALSE
#>   [118,]                        FALSE                        FALSE
#>   [119,]                        FALSE                        FALSE
#>   [120,]                        FALSE                        FALSE
#>   [121,]                        FALSE                        FALSE
#>   [122,]                        FALSE                        FALSE
#>   [123,]                        FALSE                        FALSE
#>   [124,]                        FALSE                        FALSE
#>   [125,]                        FALSE                        FALSE
#>   [126,]                        FALSE                        FALSE
#>   [127,]                        FALSE                        FALSE
#>   [128,]                        FALSE                        FALSE
#>   [129,]                        FALSE                        FALSE
#>   [130,]                        FALSE                        FALSE
#>   [131,]                        FALSE                        FALSE
#>   [132,]                        FALSE                        FALSE
#>   [133,]                        FALSE                        FALSE
#>   [134,]                        FALSE                        FALSE
#>   [135,]                        FALSE                        FALSE
#>   [136,]                        FALSE                        FALSE
#>   [137,]                        FALSE                        FALSE
#>   [138,]                        FALSE                        FALSE
#>   [139,]                        FALSE                        FALSE
#>   [140,]                        FALSE                        FALSE
#>   [141,]                        FALSE                        FALSE
#>   [142,]                        FALSE                        FALSE
#>   [143,]                        FALSE                        FALSE
#>   [144,]                        FALSE                        FALSE
#>   [145,]                         TRUE                         TRUE
#>   [146,]                        FALSE                        FALSE
#>   [147,]                         TRUE                         TRUE
#>   [148,]                        FALSE                        FALSE
#>   [149,]                        FALSE                        FALSE
#>   [150,]                        FALSE                        FALSE
#>   [151,]                        FALSE                        FALSE
#>   [152,]                        FALSE                        FALSE
#>   [153,]                        FALSE                        FALSE
#>   [154,]                        FALSE                        FALSE
#>   [155,]                        FALSE                        FALSE
#>   [156,]                        FALSE                        FALSE
#>   [157,]                        FALSE                        FALSE
#>   [158,]                        FALSE                        FALSE
#>   [159,]                        FALSE                        FALSE
#>   [160,]                        FALSE                        FALSE
#>   [161,]                        FALSE                        FALSE
#>   [162,]                        FALSE                        FALSE
#>   [163,]                        FALSE                        FALSE
#>   [164,]                        FALSE                        FALSE
#>   [165,]                        FALSE                        FALSE
#>   [166,]                        FALSE                        FALSE
#>  [ getOption("max.print") est atteint -- 13753 lignes omises ]