Computes the number of proteins that are only defined by specific peptides, shared peptides or a mixture of two.

This function computes few values about the adjacency matrix such as the number of proteins that are only defined by specific peptides, shared peptides or a mixture of two.

matAdjStats(X)

Arguments

X
The adjacency matrix with both specific and shared peptides.

Value

A list of values: * nbPeptides: the number of peptides in the matrix, nbSpecificPeptides: the number of specific peptides in the matrix, nbSharedPeptides: the number of shared peptides in the matrix, nbProt: the number of proteins in the matrix, protOnlyUniquePep: the list of proteins only defined by specific peptides, protOnlySharedPep: the list of proteins only defined by shared peptides, protMixPep: the list of proteins defined by both shared and specific peptides.

Author

Samuel Wieczorek

Examples

utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]], 'all')
matAdjStats(X)

#> $nbPeptides
#> [1] 1000
#> 
#> $nbSpecificPeptides
#> [1] 967
#> 
#> $nbSharedPeptides
#> [1] 33
#> 
#> $nbProt
#> [1] 670
#> 
#> $protOnlyUniquePep
#>   [1] "1005" "1008" "1015" "1017" "102"  "1024" "103"  "1033" "1035" "1037"
#>  [11] "104"  "1042" "1046" "105"  "1052" "1055" "1058" "106"  "1060" "1063"
#>  [21] "1064" "107"  "1073" "1074" "1077" "1079" "1080" "1086" "1088" "1092"
#>  [31] "1094" "1098" "1099" "1100" "1101" "1102" "1107" "1109" "111"  "1112"
#>  [41] "1114" "1120" "1127" "1129" "113"  "1134" "1136" "1142" "1161" "1163"
#>  [51] "1167" "1168" "1169" "1171" "1172" "118"  "1185" "1187" "1189" "1191"
#>  [61] "1192" "1198" "120"  "1202" "1206" "1207" "1208" "121"  "1210" "1212"
#>  [71] "1214" "1216" "1218" "122"  "1220" "1225" "1226" "1230" "1238" "1239"
#>  [81] "1245" "1246" "1247" "1251" "1255" "1257" "1261" "1262" "1268" "1277"
#>  [91] "1280" "1287" "1288" "129"  "1292" "1296" "1301" "1311" "1320" "1324"
#> [101] "1334" "1337" "1344" "1348" "1349" "1351" "1356" "1368" "1374" "1384"
#> [111] "1385" "1392" "1399" "1400" "1405" "1406" "1407" "1416" "1419" "1421"
#> [121] "1429" "143"  "1430" "1431" "1433" "1435" "1437" "1445" "1449" "1452"
#> [131] "1457" "146"  "1461" "1462" "1463" "1469" "1471" "1475" "1478" "1480"
#> [141] "149"  "1493" "15"   "150"  "1506" "1510" "1519" "152"  "1523" "1532"
#> [151] "1536" "1548" "155"  "1551" "1553" "1556" "1557" "156"  "1562" "1563"
#> [161] "1566" "157"  "1570" "1574" "1582" "159"  "1600" "1602" "1608" "1618"
#> [171] "1622" "1623" "164"  "1645" "1646" "1650" "1651" "1652" "1655" "1656"
#> [181] "1657" "1659" "1660" "1671" "1672" "1674" "1694" "17"   "1705" "1708"
#> [191] "1713" "1717" "1721" "1723" "173"  "1731" "1732" "1733" "174"  "1743"
#> [201] "1745" "1747" "1748" "1756" "1760" "1762" "1768" "1769" "1778" "1781"
#> [211] "1783" "1787" "1788" "1791" "1798" "18"   "180"  "1802" "1806" "181" 
#> [221] "1810" "1811" "1812" "1815" "1819" "182"  "1820" "1825" "1826" "1827"
#> [231] "183"  "1838" "184"  "1840" "1845" "185"  "1855" "1859" "186"  "1860"
#> [241] "1865" "1866" "1878" "1880" "1884" "1886" "1892" "1908" "1911" "1913"
#> [251] "1916" "1919" "1920" "1922" "1930" "1934" "1936" "1938" "1941" "1943"
#> [261] "1944" "1945" "1946" "195"  "1954" "1958" "1959" "1961" "1965" "1969"
#> [271] "1974" "1977" "198"  "1981" "1982" "1987" "199"  "1991" "1994" "1995"
#> [281] "1996" "2"    "2004" "2006" "2008" "201"  "2010" "2011" "2012" "2019"
#> [291] "202"  "2020" "2028" "2038" "204"  "206"  "2067" "207"  "2070" "2077"
#> [301] "2081" "2088" "2100" "2103" "2105" "2112" "2114" "212"  "2127" "2139"
#> [311] "214"  "2144" "2147" "2158" "216"  "2165" "2169" "217"  "2173" "2179"
#> [321] "2186" "2188" "2193" "2198" "2199" "220"  "2204" "2206" "2207" "2209"
#> [331] "2219" "2221" "2222" "2223" "2229" "223"  "2230" "2231" "2232" "2239"
#> [341] "2245" "2248" "2249" "225"  "2250" "2253" "2257" "2258" "2263" "2267"
#> [351] "228"  "2282" "2284" "2287" "2292" "2297" "230"  "2306" "2308" "2309"
#> [361] "231"  "2312" "2316" "2318" "2320" "2326" "2327" "2336" "2337" "2339"
#> [371] "2340" "2341" "2343" "2347" "235"  "236"  "237"  "2374" "2379" "238" 
#> [381] "239"  "240"  "241"  "243"  "244"  "246"  "250"  "254"  "261"  "262" 
#> [391] "263"  "265"  "272"  "274"  "275"  "276"  "277"  "279"  "28"   "280" 
#> [401] "282"  "285"  "288"  "29"   "292"  "293"  "299"  "3"    "30"   "300" 
#> [411] "304"  "309"  "311"  "314"  "317"  "32"   "320"  "321"  "324"  "325" 
#> [421] "326"  "330"  "335"  "34"   "346"  "35"   "352"  "358"  "359"  "36"  
#> [431] "360"  "361"  "364"  "366"  "368"  "370"  "376"  "377"  "378"  "379" 
#> [441] "383"  "385"  "386"  "388"  "391"  "392"  "394"  "395"  "396"  "397" 
#> [451] "40"   "400"  "401"  "408"  "410"  "411"  "413"  "414"  "42"   "423" 
#> [461] "426"  "428"  "431"  "433"  "439"  "440"  "447"  "448"  "450"  "451" 
#> [471] "452"  "457"  "459"  "460"  "463"  "464"  "465"  "467"  "471"  "472" 
#> [481] "473"  "474"  "478"  "479"  "480"  "484"  "485"  "488"  "491"  "496" 
#> [491] "497"  "498"  "50"   "500"  "505"  "515"  "518"  "520"  "524"  "527" 
#> [501] "530"  "534"  "539"  "54"   "547"  "55"   "558"  "563"  "569"  "57"  
#> [511] "573"  "574"  "576"  "58"   "582"  "597"  "598"  "599"  "601"  "604" 
#> [521] "605"  "606"  "607"  "608"  "609"  "610"  "612"  "615"  "618"  "62"  
#> [531] "620"  "63"   "631"  "644"  "653"  "655"  "657"  "658"  "660"  "661" 
#> [541] "662"  "665"  "672"  "673"  "676"  "678"  "683"  "687"  "690"  "696" 
#> [551] "704"  "706"  "708"  "711"  "723"  "725"  "728"  "732"  "734"  "735" 
#> [561] "736"  "737"  "740"  "742"  "746"  "748"  "750"  "751"  "753"  "755" 
#> [571] "756"  "757"  "758"  "762"  "763"  "766"  "767"  "768"  "778"  "779" 
#> [581] "781"  "783"  "788"  "791"  "796"  "800"  "805"  "807"  "821"  "822" 
#> [591] "824"  "826"  "828"  "829"  "834"  "838"  "84"   "841"  "843"  "845" 
#> [601] "846"  "85"   "851"  "858"  "859"  "862"  "865"  "872"  "88"   "881" 
#> [611] "889"  "905"  "908"  "910"  "914"  "933"  "935"  "94"   "946"  "95"  
#> [621] "954"  "958"  "973"  "976"  "977"  "982"  "984"  "988" 
#> 
#> $protOnlySharedPep
#>  [1] "1121" "115"  "116"  "1304" "1315" "137"  "145"  "1569" "1585" "1612"
#> [11] "167"  "1782" "1847" "2233" "2342" "245"  "295"  "302"  "306"  "490" 
#> [21] "532"  "691"  "869"  "944" 
#> 
#> $protMixPep
#>  [1] "1082" "1254" "132"  "1577" "1578" "1685" "210"  "2260" "2349" "251" 
#> [11] "252"  "257"  "289"  "290"  "362"  "717"  "986"  "987" 
#>