Generator of simulated values

Generator of simulated values

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n
An integer which is the number of simulation (same as in rbeta)

min
An integer that corresponds to the lower bound of the interval

max
An integer that corresponds to the upper bound of the interval

param1
An integer that is the first parameter of rbeta function.

param2
An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)

#>    [1] 8.769440 7.037699 9.209828 8.733655 6.676039 9.725746 8.660051 5.250345
#>    [9] 6.713652 6.884006 7.249858 6.985467 8.055310 7.895712 6.100635 9.813276
#>   [17] 6.279767 8.755952 9.870048 8.223745 9.940413 9.744668 8.364297 7.665487
#>   [25] 7.385309 9.277023 9.795163 6.117713 6.446988 7.358598 9.645806 5.812016
#>   [33] 5.452968 5.327788 7.984422 8.338143 5.687400 5.874920 6.345557 9.491307
#>   [41] 6.706009 5.193095 7.238734 9.057299 9.765646 7.183411 6.563880 5.115964
#>   [49] 8.127046 8.341072 8.647961 5.639272 7.761629 7.126459 6.845494 6.052611
#>   [57] 5.541902 9.120477 9.861945 5.647081 9.586853 7.476816 6.490248 7.779814
#>   [65] 7.983158 5.875414 7.427869 8.064944 5.251642 8.659402 9.658654 8.857072
#>   [73] 8.025112 9.731842 6.454101 8.961154 7.753874 5.098074 5.618906 8.404561
#>   [81] 8.186813 5.407627 9.592364 9.578558 6.312749 6.541570 7.905779 9.288387
#>   [89] 6.118486 5.030335 8.365585 5.823763 7.816371 6.998531 7.856827 7.115033
#>   [97] 6.979939 9.427698 5.726390 6.722605 5.690558 8.503655 6.734197 5.920231
#>  [105] 7.077911 7.549191 5.407893 9.842643 8.237792 5.057750 9.400488 5.168233
#>  [113] 6.367264 5.205189 6.885480 6.388721 7.347038 5.762993 8.219123 9.101281
#>  [121] 7.494322 6.862431 7.719644 8.986260 6.748271 8.930274 5.722074 5.422942
#>  [129] 5.429728 6.786699 6.692100 7.494332 5.604860 8.120397 6.851697 5.500588
#>  [137] 5.047662 5.372920 7.716070 8.349663 6.383641 8.931042 6.974569 5.913221
#>  [145] 5.281967 9.923369 5.197835 9.022165 6.244443 7.467731 7.588287 9.227957
#>  [153] 6.345315 5.972018 6.792036 7.311937 9.371191 9.210763 8.767531 9.166171
#>  [161] 9.635104 8.431974 7.666632 8.039932 7.082406 5.413992 5.061965 5.801221
#>  [169] 5.790914 9.550888 7.250216 6.203399 5.263250 5.287982 9.388553 6.833334
#>  [177] 6.315682 5.430435 9.719249 8.093649 5.458833 7.014473 9.980505 7.308303
#>  [185] 8.886243 7.138765 6.675228 8.763796 8.875326 6.516512 8.718456 6.978635
#>  [193] 8.143598 6.421099 9.995673 6.865424 5.817816 8.574637 7.894923 9.716773
#>  [201] 8.079265 7.763121 7.709411 7.078244 6.782602 8.173407 6.050008 8.721209
#>  [209] 7.367446 6.373567 8.995965 8.724340 8.698209 6.478324 6.507633 8.113481
#>  [217] 7.165174 6.935156 6.831202 6.968815 9.358643 7.969811 5.856558 7.418815
#>  [225] 7.415888 6.742790 6.150158 8.943123 9.113013 6.518853 6.001521 5.220219
#>  [233] 9.645216 8.452190 6.514661 5.327537 8.073475 7.340853 6.249543 9.052258
#>  [241] 7.238305 5.655620 8.370714 8.501045 8.967832 7.527974 7.711904 6.099423
#>  [249] 8.860023 9.034807 5.794790 8.073303 7.221271 5.765991 5.967362 8.516022
#>  [257] 7.139434 8.453721 8.773969 9.756325 6.227016 6.275136 6.976096 9.073671
#>  [265] 5.994025 6.153714 6.486824 7.741740 7.091354 9.150606 8.725911 8.467620
#>  [273] 9.119372 5.734972 9.609438 8.541075 8.750700 6.587508 8.732571 5.327093
#>  [281] 7.216651 6.248551 8.602125 8.168579 5.638513 8.224034 6.061601 8.038531
#>  [289] 9.192728 9.558835 6.615471 7.876518 8.982573 9.517311 5.463895 9.535156
#>  [297] 6.051320 5.849548 7.298995 6.246369 6.074759 9.772545 5.318843 6.050617
#>  [305] 8.052002 9.454186 6.143528 7.085483 9.319628 6.776143 5.335134 8.802586
#>  [313] 9.681005 9.592007 7.362505 7.030694 5.947088 9.627576 6.975613 5.525201
#>  [321] 9.963099 5.907121 7.535684 7.597267 6.575904 9.463447 7.032611 8.746805
#>  [329] 7.662142 8.833792 8.741143 5.047460 9.984692 5.445969 9.908403 7.173690
#>  [337] 9.135374 8.889947 8.112732 6.191762 6.865078 9.771155 7.408400 6.301977
#>  [345] 8.458019 9.951743 6.248972 5.106071 5.993793 8.522943 7.797642 6.666073
#>  [353] 8.669367 8.325764 8.828003 9.105397 7.927326 7.168728 8.666938 9.343007
#>  [361] 9.469549 8.353374 9.105804 8.726415 7.966667 5.420463 9.530872 6.894501
#>  [369] 9.915740 6.156050 7.546979 6.289703 9.109311 5.075339 9.764870 7.395227
#>  [377] 8.320147 7.520934 7.294051 9.388761 8.255236 8.853042 9.741858 5.366287
#>  [385] 8.225454 9.982808 7.132187 9.976745 7.410671 6.948289 5.987470 7.609648
#>  [393] 8.810720 5.014601 7.977418 9.079836 9.802893 6.619056 8.892499 7.725321
#>  [401] 6.910196 9.103066 6.542798 9.409756 6.015395 8.040847 9.092604 6.196621
#>  [409] 7.129395 6.670927 8.468370 5.382347 6.527741 8.933838 9.149527 7.908512
#>  [417] 9.445742 9.914581 6.398339 5.848533 6.194622 8.638290 8.097669 8.296220
#>  [425] 5.712670 7.122173 9.865931 6.196568 9.741458 6.281295 7.220149 7.715402
#>  [433] 9.981367 7.284102 8.278197 7.347383 5.202735 8.023538 5.169885 6.871555
#>  [441] 7.597442 8.700811 7.554345 8.189382 6.673774 5.076621 8.599497 7.653281
#>  [449] 9.496087 7.537871 6.090014 7.048456 6.962903 6.928140 6.589161 5.331719
#>  [457] 8.535115 6.099474 6.790383 6.327118 7.854817 9.633379 8.625826 6.860369
#>  [465] 9.410479 6.967535 6.520431 7.206711 9.886067 8.903372 8.166012 8.888723
#>  [473] 5.441178 9.823351 7.375685 8.750471 9.082428 7.161862 8.215035 8.158600
#>  [481] 8.927790 8.334923 6.730794 7.013619 9.240343 7.340727 8.005675 7.704007
#>  [489] 6.739979 9.917161 9.569498 5.105276 7.582670 9.174304 6.983213 5.503849
#>  [497] 8.009214 7.506344 8.855919 7.585358 8.161026 9.116187 5.032189 5.865773
#>  [505] 5.460470 6.495392 5.197552 6.618443 6.127108 5.543562 8.558752 7.882386
#>  [513] 5.348149 7.431981 6.920635 9.490115 7.216714 6.546950 6.988966 8.147941
#>  [521] 9.854439 5.735117 9.664334 7.204319 8.944075 7.188516 5.585029 9.679854
#>  [529] 8.634983 8.502125 8.148337 6.255459 8.736728 6.493508 5.407137 5.712251
#>  [537] 7.717788 9.580399 7.798978 9.460595 7.213154 5.236927 6.863759 9.640888
#>  [545] 5.894402 6.477187 6.176429 6.666064 8.055790 6.190941 9.555740 8.801187
#>  [553] 5.787269 5.800566 9.467689 6.131523 7.788435 7.895576 6.129852 6.936137
#>  [561] 7.212738 9.352041 5.803852 8.438251 8.604783 8.850518 9.262567 6.023494
#>  [569] 5.049653 8.670190 8.235749 5.042241 9.129719 6.208916 6.387003 6.549636
#>  [577] 6.975429 5.850921 9.256035 6.326562 8.234889 8.149155 9.573927 5.956362
#>  [585] 8.686943 7.873535 8.122960 8.917925 9.564036 5.395830 6.833910 7.383960
#>  [593] 9.949166 6.327698 6.970798 8.291931 6.599998 6.702501 6.844956 8.534187
#>  [601] 8.860459 9.555761 9.458375 8.670569 5.870613 6.745798 9.655007 7.343592
#>  [609] 9.173029 9.000584 9.732398 8.822495 5.617621 5.325182 5.993003 8.153426
#>  [617] 5.650844 9.653181 9.112802 7.736605 5.402056 8.051286 6.929824 7.068642
#>  [625] 8.543029 9.265434 5.798641 5.006407 5.446472 6.108129 7.686134 5.362054
#>  [633] 7.492343 6.611998 8.139416 5.397998 8.588263 9.777386 8.786264 9.723166
#>  [641] 6.924999 9.169895 6.125236 9.152486 9.615450 6.812064 5.761136 8.227260
#>  [649] 6.978732 5.430034 8.382306 9.259464 6.981508 6.749102 5.576553 8.347443
#>  [657] 6.077503 6.922669 7.214297 5.685822 7.356394 7.319755 6.920279 9.577309
#>  [665] 9.281336 8.346983 5.058252 7.810369 5.086561 9.693888 7.310859 7.042175
#>  [673] 5.642999 5.763203 8.534214 5.129594 5.355730 7.931543 8.072031 5.820442
#>  [681] 5.821879 5.845554 5.889072 9.747593 7.400604 9.547164 7.932636 6.601037
#>  [689] 5.617183 8.820417 9.781236 9.711944 7.260221 5.848928 8.811132 9.282066
#>  [697] 8.374899 7.101526 5.565884 8.597633 5.873331 8.552494 7.895513 8.040971
#>  [705] 6.540179 5.744180 7.805316 5.404951 9.885512 5.030823 8.644547 9.881687
#>  [713] 7.512827 7.902586 6.956782 6.341562 6.467891 8.631333 7.565995 8.465727
#>  [721] 7.666178 5.612551 5.418079 5.386663 5.601475 5.985241 8.914472 9.709368
#>  [729] 9.428990 9.030773 8.899174 9.767529 9.518071 9.327889 9.524410 8.665656
#>  [737] 9.878618 9.469978 8.457752 7.141569 9.056246 6.010185 8.633379 7.161629
#>  [745] 7.464391 9.671913 8.364783 9.850237 5.791117 7.942406 7.362910 9.664444
#>  [753] 9.251436 8.103139 7.101818 6.192128 9.701275 8.703070 9.265503 9.730344
#>  [761] 9.431044 9.751491 5.673640 9.842094 9.943710 8.558650 8.504123 5.022262
#>  [769] 6.118938 9.667960 6.856616 9.371303 8.970934 5.756057 7.739930 6.082957
#>  [777] 6.272535 7.593508 5.803538 5.170271 8.423320 5.212174 7.294961 7.745263
#>  [785] 8.075560 9.641546 8.163271 6.931518 7.124109 5.311995 9.364419 6.864179
#>  [793] 8.966744 8.402782 9.857859 6.964811 7.244339 5.704343 8.565943 7.002657
#>  [801] 7.458139 7.050005 9.541031 6.883560 7.489535 8.851836 7.150871 7.530614
#>  [809] 7.865119 9.930985 8.496633 9.981741 5.958258 5.925685 7.287675 9.467516
#>  [817] 8.559113 5.114728 6.508135 8.101223 8.827220 9.646148 8.655877 9.389371
#>  [825] 5.613603 7.048446 6.393279 9.772480 5.760638 9.662609 9.100908 5.063630
#>  [833] 6.387373 6.366682 9.777364 5.509448 9.481670 8.783733 8.830571 5.410552
#>  [841] 8.073379 5.769194 5.488941 9.107094 8.327916 6.196600 7.657487 7.986544
#>  [849] 8.200357 9.443903 6.166964 5.760876 9.190348 7.480607 6.169022 7.105385
#>  [857] 6.764349 5.637463 7.393125 7.541396 8.090008 8.824679 6.644554 9.946906
#>  [865] 7.589152 8.593698 8.614113 6.440505 6.680890 8.561591 8.427869 6.730295
#>  [873] 9.889858 8.362320 7.460155 5.289916 8.959285 6.618316 5.825000 7.574572
#>  [881] 7.356478 9.393179 7.655691 5.607663 5.273392 8.243671 9.385348 5.818709
#>  [889] 6.050698 9.350255 9.701684 6.144120 5.330878 5.753816 9.447976 8.289034
#>  [897] 5.903583 5.758111 9.641518 5.164571 5.185504 9.757132 9.641210 7.607122
#>  [905] 9.581063 6.285442 8.829440 8.925291 6.688702 8.219068 5.543923 7.617650
#>  [913] 5.730012 9.368638 6.038112 8.229016 7.448935 6.214411 7.130390 8.835056
#>  [921] 8.853055 9.212562 9.805128 8.914265 8.572001 6.034191 9.434768 6.462594
#>  [929] 5.270303 5.042363 9.173438 9.215159 8.550151 6.985186 5.490409 6.667233
#>  [937] 6.199009 5.231627 6.776359 8.678179 7.131859 7.775345 9.320639 9.883000
#>  [945] 9.378951 7.267910 7.949346 6.690453 6.823558 5.925034 6.737133 7.743566
#>  [953] 9.044328 5.697586 7.431312 7.486621 7.302530 7.966727 6.568890 8.661496
#>  [961] 6.090822 8.795451 8.854292 9.313604 8.933798 9.010795 8.879462 6.256215
#>  [969] 9.108497 6.543682 6.476784 8.586174 5.175138 9.359940 9.819185 6.091018
#>  [977] 5.221119 9.893016 6.271194 5.270683 6.598136 7.445150 9.579068 8.830241
#>  [985] 9.967485 6.735567 6.049590 6.929371 7.633194 8.015296 5.774994 6.501318
#>  [993] 6.989174 5.613386 9.064642 7.768103 7.671933 8.482764 9.811672 6.674398