Method to aggregate peptides into proteins with the iterative approach.

Method to aggregate peptides into proteins with the iterative approach

aggIter(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)

Arguments

qPepData
A matrix of intensities of peptides.

X
An adjacency matrix

conditions
xxx

init.method
The method used to initialize the iterative algotirhm. Default is 'Sum'.

method
The method used for xxx. Default value is 'Mean'.

n
xxxx

Value

A matrix of protein intensities.

Author

Samuel Wieczorek

Examples

library(QFeatures)

#> Le chargement a nécessité le package : MultiAssayExperiment

#> Le chargement a nécessité le package : SummarizedExperiment

#> Warning: le package 'SummarizedExperiment' a été compilé avec la version R 4.0.2

#> Le chargement a nécessité le package : GenomicRanges

#> Le chargement a nécessité le package : stats4

#> Le chargement a nécessité le package : BiocGenerics

#> Le chargement a nécessité le package : parallel

#> 
#> Attachement du package : 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Le chargement a nécessité le package : S4Vectors

#> 
#> Attachement du package : 'S4Vectors'

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> Le chargement a nécessité le package : IRanges

#> 
#> Attachement du package : 'IRanges'

#> The following object is masked from 'package:grDevices':
#> 
#>     windows

#> Le chargement a nécessité le package : GenomeInfoDb

#> Warning: le package 'GenomeInfoDb' a été compilé avec la version R 4.0.2

#> Le chargement a nécessité le package : Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Le chargement a nécessité le package : DelayedArray

#> Warning: le package 'DelayedArray' a été compilé avec la version R 4.0.2

#> Le chargement a nécessité le package : Matrix

#> 
#> Attachement du package : 'Matrix'

#> The following object is masked from 'package:S4Vectors':
#> 
#>     expand

#> Le chargement a nécessité le package : MatrixGenerics

#> Warning: le package 'MatrixGenerics' a été compilé avec la version R 4.0.2

#> Le chargement a nécessité le package : matrixStats

#> Warning: le package 'matrixStats' a été compilé avec la version R 4.0.2

#> 
#> Attachement du package : 'matrixStats'

#> The following objects are masked from 'package:Biobase':
#> 
#>     anyMissing, rowMedians

#> 
#> Attachement du package : 'MatrixGenerics'

#> The following objects are masked from 'package:matrixStats':
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars

#> The following object is masked from 'package:Biobase':
#> 
#>     rowMedians

#> 
#> Attachement du package : 'DelayedArray'

#> The following objects are masked from 'package:base':
#> 
#>     aperm, apply, rowsum

#> 
#> Attachement du package : 'QFeatures'

#> The following object is masked from 'package:DelayedArray':
#> 
#>     sweep

#> The following object is masked from 'package:base':
#> 
#>     sweep

utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]], 'all'))
conditions <- colData(obj)$Condition
aggIter(assay(obj,2), X, conditions)

#>            [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#>   [1,] 24.28821 24.48162 24.32201 24.10576 24.28980 25.45394
#>   [2,] 24.23783 23.38166      NaN      NaN      NaN      NaN
#>   [3,]      NaN      NaN      NaN      NaN      NaN      NaN
#>   [4,] 22.50959 22.67980 22.93771 22.55190 22.71791 22.58516
#>   [5,] 26.09462 26.21429 26.16749 26.09004 26.01065 26.09375
#>   [6,] 24.17805 24.10837 23.74755 24.13438 24.88646 24.02547
#>   [7,] 21.92136 21.55936      NaN 22.78337 22.34627      NaN
#>   [8,] 23.19862 22.03234 23.18264 22.87017 22.59332 22.85274
#>   [9,] 27.73982 27.63242 27.89151 27.59328 27.68826 27.73286
#>  [10,] 22.01828 22.31921 22.13877 22.46914 22.61715 22.52920
#>  [11,]      NaN      NaN 21.70092      NaN      NaN      NaN
#>  [12,] 22.62053 23.07604 23.03971 23.12210 22.78965 23.08937
#>  [13,] 26.04675 26.16701 26.16127 26.09831 26.08242 26.03132
#>  [14,] 30.34763 30.31490 30.43516 30.38733 30.38177 30.35366
#>  [15,] 23.12366 23.25507 23.25249 23.32313 23.07801 23.35860
#>  [16,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [17,] 23.91031 23.95136 23.77967 24.12305 23.96455 23.86703
#>  [18,] 29.87954 29.89272 30.03522 30.40243 30.29909 30.42482
#>  [19,] 21.31163 21.27509 22.31722 20.93049 21.14582 21.36051
#>  [20,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [21,] 23.73469 23.78726 23.98555      NaN      NaN      NaN
#>  [22,]      NaN      NaN      NaN      NaN 22.17283      NaN
#>  [23,] 23.82978 24.91452      NaN      NaN 24.56651 24.36212
#>  [24,] 25.62327 25.72882 25.62374 25.57467 25.59867 25.66909
#>  [25,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [26,] 26.03183 26.00093 26.12761 26.00755 25.96299 25.92105
#>  [27,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [28,] 24.66239 24.58734 24.82125 24.67311 24.65929 24.70313
#>  [29,]      NaN      NaN      NaN 23.50252 23.54687 23.77957
#>  [30,] 24.21076 24.17064 25.04714 24.76779 26.41700 24.47977
#>  [31,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [32,] 23.66839 23.77806 23.64240 24.33470 24.31347 24.11012
#>  [33,] 24.93463 24.87037 24.71618 24.91884 24.78905 24.80464
#>  [34,] 24.09233 23.91424 24.00273 23.92488 24.03607 24.13312
#>  [35,] 23.12307 22.98801 22.78391 22.30048 23.16987 23.14247
#>  [36,] 24.76855 24.84460 25.07997 24.80371 24.84575 24.95180
#>  [37,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [38,] 25.29399 25.27184 25.40099 25.43472 25.36470 25.15994
#>  [39,] 25.91166 25.37806 26.37814 25.56976 25.30946 25.97235
#>  [40,] 25.52925 24.96531 25.45001 24.62304 23.68222 24.56691
#>  [41,] 25.44038 24.82527 24.95916 25.05887      NaN 25.89809
#>  [42,] 28.38187 28.62697 28.24436 28.71954 28.74114 28.69962
#>  [43,] 24.30904 24.16507 24.14857 24.55462 24.05606 23.84048
#>  [44,] 29.00603 28.83993 29.03708      NaN      NaN      NaN
#>  [45,] 22.74411 23.07084 23.43288 23.80863 23.67347 23.63998
#>  [46,] 22.10856 22.91929 22.77429      NaN 23.45237 22.30265
#>  [47,] 23.50736 23.68560 23.59149 23.67691 23.53505 23.56674
#>  [48,] 26.39432 26.38744 25.14383 27.04689 26.70654 25.93712
#>  [49,] 23.81685 23.57946 23.61392 23.55862 23.44594 23.66618
#>  [50,] 21.76300 22.04634 22.09226 22.42299 22.52900 22.35539
#>  [51,] 31.83913 31.83195 31.89537 31.64074 31.64583 31.68345
#>  [52,] 31.19671 31.08480 31.26153 31.64551 31.65057 31.68774
#>  [53,] 22.92982 22.89145 22.94163 22.97733 23.32237 22.79634
#>  [54,] 23.11475 22.97586 23.01556 23.64329 23.11671 23.14863
#>  [55,] 23.55508 22.69379      NaN 22.64834 22.84059 22.70440
#>  [56,] 23.17910 23.01995 23.20964 23.08814 23.09396 23.16723
#>  [57,] 22.33415 22.77133 23.89125 23.61941 22.56656 23.33347
#>  [58,]      NaN      NaN      NaN 22.24925 22.92276      NaN
#>  [59,] 23.30953 22.41460      NaN 23.34570 22.88506      NaN
#>  [60,] 22.98155 22.83123 23.10402 23.10018 22.90452 23.07612
#>  [61,] 25.63533 25.63056 25.57016 25.46927 25.71837 26.00041
#>  [62,]      NaN      NaN      NaN      NaN 21.50688 21.79713
#>  [63,] 24.78816 24.75364 24.89947 24.68587 24.76228 24.73593
#>  [64,] 23.36946 22.41970 23.36973 24.29740 24.20527 24.42865
#>  [65,] 24.22724 24.48470 23.99575 24.65502 25.11474 24.91401
#>  [66,]      NaN      NaN 26.66302      NaN      NaN      NaN
#>  [67,] 29.33685 29.23696 29.18880 28.68918 28.35326 28.43538
#>  [68,] 22.61171      NaN      NaN      NaN 22.13082 22.55929
#>  [69,] 21.38221 20.83353 22.31399 21.28416 20.66700 21.40567
#>  [70,] 23.05629 25.77733 24.28101 22.65093      NaN 23.16750
#>  [71,]      NaN      NaN      NaN 24.90495 24.88045 24.87853
#>  [72,] 29.15930 28.97954 29.17268 28.97845 28.94010 28.99323
#>  [73,] 24.81538 24.81108      NaN 25.16273 24.99593 24.48696
#>  [74,]      NaN 20.56039 23.93894      NaN      NaN      NaN
#>  [75,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [76,] 25.64078 25.98383 25.84943 26.03792 26.09563 25.89099
#>  [77,] 23.93625 24.07071 24.16645 23.79412 23.67034 24.14048
#>  [78,]      NaN      NaN      NaN 25.45951      NaN 25.69936
#>  [79,] 23.61717      NaN 23.39218 23.93457 23.81419 24.43993
#>  [80,] 25.04411 24.42284 24.80887 24.75051 26.46516 25.06218
#>  [81,]      NaN      NaN      NaN      NaN      NaN      NaN
#>  [82,] 21.01760 18.83102 21.71854 20.03308      NaN      NaN
#>  [83,]      NaN      NaN 23.78137 26.03762 25.82406 25.98372
#>  [84,] 23.01139 23.44380 23.74714 23.46501      NaN 23.98538
#>  [85,] 24.58474      NaN 24.35232 24.71382 24.60761 24.39447
#>  [86,] 25.98407 25.54371 25.70604 25.70588 25.54557 25.47988
#>  [87,] 24.14258      NaN      NaN 23.00520 22.86282 23.29824
#>  [88,] 24.22825 25.19838 25.21892 25.56023 24.54781 24.68239
#>  [89,] 25.25494 25.16158 25.25868 24.48563 24.33572 24.42122
#>  [90,] 23.96570 23.95766 23.53848 23.77485 23.70567 24.09611
#>  [91,] 22.22007 22.55314 22.46971 22.29908 22.71360 22.55880
#>  [92,] 27.07337 26.37496      NaN 26.49095 26.04415 25.82692
#>  [93,]      NaN 22.17411      NaN 22.22955      NaN      NaN
#>  [94,] 25.66668 25.56990 25.62751 25.31979 25.42172 25.37355
#>  [95,] 24.92823 24.26657 25.28207 24.37643 24.60637 24.83615
#>  [96,] 24.91315 24.96886 23.81752 23.81868 23.89989 24.94887
#>  [97,] 24.67023      NaN      NaN 25.22037      NaN 24.77369
#>  [98,] 25.60184 25.44255 25.66162 25.56721 25.35164 26.23320
#>  [99,] 27.72468 27.87121 27.91234 28.12376 27.78094 27.76945
#> [100,] 24.36466 24.49698 24.13406 24.77259 24.48370 24.51431
#> [101,] 20.94149 21.22529 20.85924 21.33642      NaN 21.01475
#> [102,] 25.21119 25.22185 25.26789 25.07521 25.01950 25.06015
#> [103,] 29.26907 29.27143 29.26901 29.34994 29.28869 29.27928
#> [104,] 26.82808 26.46610 27.14287 26.56369 26.43255 26.71463
#> [105,] 27.82771 27.63631 28.25902 27.82343 27.62708 27.93925
#> [106,] 28.64127 28.56216 28.79636 28.58788 28.62931 28.62671
#> [107,] 21.98301 22.13031 20.68526 22.06375 21.79487      NaN
#> [108,]      NaN      NaN      NaN 25.35988 25.25249 25.40720
#> [109,] 24.07423 23.87624 23.38390 23.92325 24.14950 23.85633
#> [110,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [111,]      NaN      NaN 24.80233 24.29614      NaN      NaN
#> [112,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [113,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [114,]      NaN      NaN 23.73345 25.15213 25.10274 23.95409
#> [115,] 24.02776 24.05238 24.04878 24.18867 24.20280 24.30845
#> [116,] 29.79623 29.63568 30.00437 29.71221 29.65475 29.77156
#> [117,] 26.04481 25.96420 26.00243 25.81565 25.96054 25.99252
#> [118,] 23.71691 23.44329 23.63898 23.87802 23.56941 23.80075
#> [119,] 24.05557 24.03347 23.72619 23.76850 23.73665 23.58520
#> [120,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [121,]      NaN      NaN      NaN 24.05292 24.98229 24.47673
#> [122,] 26.61470 26.58663 26.54628 26.57099 26.33451 26.58463
#> [123,] 21.85943 21.77406 21.98676 21.76036 22.29701 22.02759
#> [124,] 28.99110 28.41748 25.98748 28.74679 28.15316 27.27906
#> [125,] 24.51301 24.49550 24.84708 24.63856 24.52152 24.53274
#> [126,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [127,] 24.01683 24.27467 23.99647 24.83014 24.58450 24.91763
#> [128,] 24.04602 24.37206 24.23593 23.59537 23.46438 23.39441
#> [129,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [130,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [131,] 26.69998 26.57738 26.45916 27.16500 27.06394 27.07286
#> [132,] 23.40938      NaN 23.38298 22.30619 23.14659 22.88993
#> [133,]      NaN      NaN 21.15956      NaN      NaN 18.37581
#> [134,] 23.39414 23.20658 23.75328 23.96638 23.05552 23.66021
#> [135,]      NaN      NaN      NaN 22.19761      NaN      NaN
#> [136,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [137,] 22.03708 21.99762 22.28475 22.79677 22.46096 22.53284
#> [138,] 25.36873 25.34841 25.81490 24.63355 24.71288 24.47997
#> [139,] 30.15774 29.97856 30.04934 29.96072 29.84721 29.86616
#> [140,] 25.12589 24.87329 24.78821 25.04802 24.68094 24.81426
#> [141,] 23.47988 22.56193 23.11517 23.10795      NaN 23.77716
#> [142,] 23.88362 23.27739 23.76679 24.13140 23.69434 24.05432
#> [143,] 25.56639 25.76744 25.63156 25.58577 25.63633 25.51078
#> [144,]      NaN 24.75527 23.55039      NaN      NaN      NaN
#> [145,]      NaN      NaN      NaN 24.02724 22.92586      NaN
#> [146,] 26.91593 27.01903 26.83545 26.43212 26.33994 26.24764
#> [147,]      NaN      NaN 22.56302 23.32649 21.88651 22.88889
#> [148,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [149,] 23.79769 23.35676 22.26236 23.66557 22.68895 22.86055
#> [150,] 22.79005      NaN      NaN      NaN      NaN 21.81554
#> [151,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [152,] 23.99895 23.83759 24.12052 24.13782 24.25083 24.09925
#> [153,] 25.75580 25.93933 25.93705 24.64246 23.84929 24.20359
#> [154,] 25.47303 25.41685 25.36904 24.89979 25.47645 25.52155
#> [155,]      NaN      NaN 20.45934      NaN      NaN      NaN
#> [156,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [157,] 22.42611 22.72133 22.80323 23.08463 23.07726 23.27933
#> [158,] 25.54054 25.83918 25.57159 26.41393 26.15888 26.26835
#> [159,] 25.38647 25.10946 25.78980 25.73190 25.67331 25.55277
#> [160,] 21.28901 21.54383 21.81656 21.54727 21.76186 21.61104
#> [161,]      NaN      NaN      NaN 26.22643 26.06122 26.01342
#> [162,] 24.15971 24.06275 24.07118 24.06873 24.14158 24.34434
#> [163,] 27.86526 27.84834 27.99518 27.79017 27.79916 27.82037
#> [164,] 26.49631 26.53921 26.50511 26.53671 26.54206 26.63686
#> [165,]      NaN 22.89770 23.03922 19.83501 22.47417 23.42085
#> [166,] 25.34279      NaN      NaN      NaN      NaN 26.01704
#> [167,]      NaN 21.10527 22.11189 22.53458 22.37857 22.90185
#> [168,]      NaN      NaN      NaN 29.47983 29.51200 29.49840
#> [169,] 23.26084 22.98819 23.17841      NaN      NaN 23.44872
#> [170,] 23.99981 23.62923 23.69710 23.55801 23.47321 23.67433
#> [171,]      NaN      NaN      NaN      NaN 22.99205      NaN
#> [172,] 24.04874 23.52253 23.85609 24.06624 24.01661 23.76528
#> [173,] 30.69785 30.70891 30.83816 30.60805 30.66662 30.65224
#> [174,]      NaN      NaN 22.98848 21.15394      NaN 22.91064
#> [175,]      NaN      NaN 24.34875      NaN      NaN      NaN
#> [176,] 26.51000 26.45331 26.61121 24.64438 25.19991 25.53361
#> [177,] 25.75297 25.75642 25.74579 24.89924 24.97491 24.78217
#> [178,]      NaN      NaN      NaN 23.08451      NaN 22.90338
#> [179,] 20.47892      NaN 20.82765 20.24516 20.03389 20.66778
#> [180,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [181,] 24.53884 24.39767 24.77254 24.44607 24.62310 24.41177
#> [182,] 22.62734 22.40294 23.02059 22.58213 22.42168 22.84405
#> [183,]      NaN      NaN 21.74710 22.05351      NaN 21.83696
#> [184,] 24.75299 24.47501 24.75456 24.52173 24.55780 24.68310
#> [185,] 21.22394 22.10738 22.17454 22.47687 22.67075 22.31923
#> [186,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [187,] 23.50373 26.21477 25.55198 23.58416 23.91628 21.89048
#> [188,] 23.97508 24.09213 24.29259 24.17949 23.99054 24.04247
#> [189,] 25.45482 25.44881 25.55582 26.12562 26.23288 26.09804
#> [190,] 23.88551 23.81954 22.86280 23.50674 23.36520 23.33920
#> [191,] 21.62176 21.33134 21.73343 21.51306 23.47246 21.67814
#> [192,]      NaN      NaN      NaN 22.57953 22.96088 23.02712
#> [193,] 22.62841 22.79430 22.78997 22.87314 22.53938 22.77034
#> [194,]      NaN 22.38914      NaN 22.20932 21.93693 22.03093
#> [195,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [196,] 25.15165 24.88073 25.02925 24.57743 23.77458 24.48922
#> [197,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [198,]      NaN      NaN 20.70227 21.97694 21.56756 21.45257
#> [199,] 27.67175 27.72091 27.37832 26.90086 26.72633 26.74349
#> [200,] 26.20025 26.27035 26.26748 26.22883 26.13581 26.14552
#> [201,] 22.88258 22.98796 22.94126      NaN      NaN      NaN
#> [202,] 25.12384 25.07006 25.00217 25.32701 25.07300 24.93674
#> [203,] 26.53651 26.56981 26.65641 26.41512 26.49145 26.42880
#> [204,] 23.58703 24.32286 24.35582 23.68378 24.36252 23.90426
#> [205,] 24.28270 24.10291 24.90571 23.25240 23.16750 22.74151
#> [206,] 23.14232 22.96385 22.13659      NaN      NaN      NaN
#> [207,] 23.83297 23.57092 23.82755 23.15822 23.61672 23.78137
#> [208,] 22.26227 22.00618 22.38764 22.58724 22.46316 22.61675
#> [209,] 25.59997 25.33439 25.08310 25.19129 25.26785 25.10933
#> [210,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [211,] 21.54402 21.58311 22.16214 22.02273 21.86854 21.93557
#> [212,] 28.54504 28.68764 28.60583 28.57992 28.48969 28.53006
#> [213,] 22.53675 22.76529 22.70900 22.85890 22.41772 22.42293
#> [214,] 22.77258 22.78732 22.97848 23.08874 22.79128 22.58839
#> [215,] 24.49796 24.09668 24.29446 24.15314 24.37749 24.49460
#> [216,] 27.66309 27.73163 27.82960 27.60477 27.65983 27.65949
#> [217,] 24.34895 24.25854 24.15707      NaN      NaN      NaN
#> [218,] 22.14759 22.31188 21.85472      NaN 21.88383 22.15205
#> [219,] 24.08489 23.76476 24.20926 23.95420 23.92551 23.87511
#> [220,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [221,]      NaN      NaN      NaN 23.57980 23.63799 23.40963
#> [222,]      NaN      NaN      NaN 24.68833 24.72624 24.75623
#> [223,] 24.16223 23.31451 21.91572      NaN 23.43314 22.94138
#> [224,]      NaN      NaN      NaN      NaN 23.55602 23.54073
#> [225,] 22.61385 22.22673 23.08432 22.35089 22.03835 22.21481
#> [226,] 26.32502 26.48004 26.49445 26.46101 26.44063 26.52030
#> [227,] 21.97980      NaN 22.20538 22.24525      NaN 21.80089
#> [228,]      NaN      NaN      NaN 24.10574 23.27071 23.97202
#> [229,] 27.04945 26.96387 27.04750 26.92752 26.94911 27.04602
#> [230,] 25.39529 25.38663 25.53863 24.92325 24.87355 24.97591
#> [231,] 25.69168 25.69970 25.56349 25.74912 25.69785 25.63730
#> [232,] 21.81426 21.94535 21.98280 22.21267 21.96353 21.95865
#> [233,] 27.54078 27.48988 27.51016 27.41652 27.35563 27.42330
#> [234,] 24.60377 24.97334 24.44310 23.84029 23.27782 23.23978
#> [235,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [236,]      NaN 22.13352 22.45566      NaN      NaN 22.60356
#> [237,] 23.29460 23.07253 23.31990      NaN      NaN      NaN
#> [238,] 25.20493 24.80151 25.03929 25.27987 24.90482 24.85160
#> [239,] 23.08327 22.90584 22.95807 23.23419 23.10923 23.23181
#> [240,] 22.60498 22.76623 22.88963 22.90193 22.57902 22.97091
#> [241,] 21.17042 20.87222 20.81120      NaN 21.15653 21.43624
#> [242,] 22.71106      NaN 23.68966      NaN 22.75786 23.30786
#> [243,] 25.67250 25.68094 25.76277 25.65889 25.65770 25.58862
#> [244,] 25.80971 25.74492 26.48907 25.67130 26.25006 25.99907
#> [245,] 22.72549 23.01129 23.86872 23.34381 23.19855 24.35542
#> [246,] 23.85348 23.58095 23.68816 23.59833 23.70767 23.47333
#> [247,] 23.10066      NaN 23.97316 24.36152 24.09869 24.76157
#> [248,]      NaN      NaN      NaN 24.49607 24.11513 24.40100
#> [249,] 31.42367 31.29049 31.11942 31.74363 31.59988 31.48807
#> [250,] 23.53160      NaN      NaN 23.18666      NaN 23.38126
#> [251,] 22.04510 21.57029      NaN 21.80515 20.96670 21.00594
#> [252,]      NaN      NaN      NaN 23.70830 23.08949 23.47259
#> [253,]      NaN      NaN      NaN 24.49582 24.17820 24.29207
#> [254,] 25.66126 25.58073 25.73998 24.63484 24.64955 24.80440
#> [255,] 27.73117 27.70425 27.56660 27.77593 27.63346 27.69690
#> [256,] 22.87170      NaN      NaN 21.94935      NaN      NaN
#> [257,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [258,] 26.09806 26.04183 26.26613 26.21097 26.11709 26.34005
#> [259,]      NaN      NaN      NaN      NaN      NaN 23.38469
#> [260,] 23.12406 23.19254 23.22037 23.05792 23.11141 23.04097
#> [261,] 23.00526 23.17254 23.29660 22.82714 23.12032 23.21228
#> [262,]      NaN      NaN      NaN 23.61414 23.35703 24.46750
#> [263,] 28.79860 28.69265 28.78444 28.72537 28.60212 28.75789
#> [264,] 25.71722 25.83224 25.89088 23.64152 25.89164 26.13252
#> [265,]      NaN      NaN 22.29782 22.40491 22.06019 22.22173
#> [266,] 25.50746 24.99752 25.69490 25.06442 24.96179 25.58451
#> [267,] 21.45499 21.13907 21.71465 21.78769 21.44538 21.33358
#> [268,] 23.13641      NaN      NaN 23.13292 23.67939      NaN
#> [269,] 28.07835 28.09890 28.23777 28.06936 28.01996 28.07085
#> [270,] 23.35878 23.04413 23.66661 23.51063 23.31845 22.98786
#> [271,] 24.42470 24.58055 24.85038 24.04394 24.40743 24.48653
#> [272,] 22.69177 22.27045 22.16451 22.66830      NaN 22.37053
#> [273,] 24.85584 24.80347 24.76248 24.99835 24.64161 24.86586
#> [274,] 23.67228 23.82639 23.22613 24.20926 23.73614 23.99309
#> [275,]      NaN 22.80855 22.41658      NaN      NaN 22.55733
#> [276,] 24.77107 24.61178 24.72250 24.41041 24.68014 24.91812
#> [277,] 25.58580 25.61555 25.50306 25.56532 25.53623 25.52692
#> [278,]      NaN      NaN      NaN 25.19423 25.29267 25.40009
#> [279,] 25.50879 25.53452 25.69046 25.54713 25.63082 25.50126
#> [280,]      NaN 22.55579 22.45126      NaN 22.86461      NaN
#> [281,] 25.50130 25.35262 25.39875 25.16223 25.47510 25.25032
#> [282,] 22.87917 23.10271 23.68878 23.11004 23.01820 23.13935
#> [283,] 23.41622 22.85802 22.60168 22.95851 23.43860 23.42816
#> [284,] 27.23827 27.13055 27.23444 27.08669 27.24145 27.11486
#> [285,] 24.51492 24.38680 24.61957 24.51710 24.46588 24.37231
#> [286,] 22.14127 22.44660 22.62807 22.19060 22.27665 22.46044
#> [287,] 23.47222      NaN 23.29712      NaN      NaN      NaN
#> [288,] 24.33851 24.18790 24.16254 24.01052 24.15561 24.26420
#> [289,] 24.57646 24.60264 24.54339 24.53925 24.32237 24.90467
#> [290,]      NaN 22.86439 22.53467      NaN 23.10161 23.15442
#> [291,] 23.76598 23.53303 23.96199 23.72910 23.66915 24.07757
#> [292,]      NaN      NaN      NaN      NaN 22.11572      NaN
#> [293,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [294,] 24.52981 24.60105 24.60919 24.13938 24.51929 24.51328
#> [295,] 24.37987 24.49467 24.51983 24.68367 24.30243 24.25508
#> [296,] 26.38358 26.20166 26.62162 26.24191 26.15954 26.26666
#> [297,]      NaN      NaN 23.93562 23.21410 23.76304      NaN
#> [298,]      NaN      NaN      NaN      NaN      NaN 22.47070
#> [299,] 24.59315 24.14266 24.30508 24.42899 25.92309 25.81370
#> [300,] 25.09434 24.77279 24.74289 24.65824 24.72057 24.86712
#> [301,] 25.94505 25.43755 26.48222 24.65078 24.35595 24.77554
#> [302,] 25.15221 24.90440 25.13739 25.04323 24.74006 23.99430
#> [303,] 23.83576 23.30368 23.42406 23.35716 23.27952 23.54156
#> [304,]      NaN 23.39467      NaN 23.29134 23.97719 24.11712
#> [305,]      NaN      NaN 23.23989      NaN      NaN 23.01759
#> [306,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [307,] 21.79994 22.12142 21.79205      NaN 21.50378 21.69574
#> [308,] 23.32416 23.73148 23.55227 23.27810 23.61784 23.54038
#> [309,] 24.86637 24.56703 24.47029 24.59895 24.74919 24.90738
#> [310,] 24.39983 24.73841 24.42425 23.48499 23.47046 23.40036
#> [311,] 25.42463 25.20945 25.52577 25.71888 25.15334 25.66182
#> [312,] 26.84074 26.82905 26.83425 26.72312 26.59021 26.66871
#> [313,] 23.36159 23.38956      NaN      NaN 24.19895 24.21801
#> [314,] 23.90919 22.91494 24.38863 25.27583 24.24937 23.88698
#> [315,] 26.75162 26.69697 26.78294 26.72429 26.75353 26.70425
#> [316,] 24.20598 23.96393 24.37994 24.33497 24.07006 24.29418
#> [317,] 23.62288 23.78915 23.76182 23.54898 23.48800 23.76436
#> [318,]      NaN      NaN      NaN 22.77241 23.05762 22.40624
#> [319,] 23.84479 23.79769 24.05937 24.23118 24.00625 24.06521
#> [320,]      NaN      NaN      NaN 23.81851 23.64933 23.77304
#> [321,] 22.94192 22.81355 22.91676 23.51196 23.12207 23.65666
#> [322,] 23.53206 23.63810 23.67204 23.63219 23.71851 23.73977
#> [323,] 23.16518      NaN 22.65424      NaN 23.55637      NaN
#> [324,]      NaN 20.99797 21.57980      NaN      NaN      NaN
#> [325,] 23.25969 23.52061 23.38443 23.11119 23.19092 23.16066
#> [326,] 25.61593 25.51389 25.81236 25.37657 25.32166 25.54141
#> [327,]      NaN      NaN      NaN 23.11264      NaN      NaN
#> [328,] 25.28510 25.51701 26.17627 25.52645 22.92013 25.11291
#> [329,]      NaN      NaN      NaN 22.30627 22.21824 22.16576
#> [330,]      NaN      NaN      NaN      NaN 21.91412      NaN
#> [331,] 25.72385 25.57437 25.73749 25.63983 25.60670 25.60708
#> [332,] 24.16806 24.16392 24.18170 24.07030 24.14429 24.14911
#> [333,] 23.48800 23.38007 23.26127 23.53671 22.93229 23.18595
#> [334,] 23.54392 23.51051      NaN      NaN      NaN      NaN
#> [335,] 24.59235 24.41654 24.68319 24.52689 24.49954 24.65009
#> [336,] 24.96512 23.74066 24.89878 24.94699 23.73579 23.57108
#> [337,] 21.99293      NaN 22.37917      NaN 22.78955 22.10140
#> [338,] 24.14600 23.52910      NaN 24.02166 23.01952      NaN
#> [339,] 22.06819 22.20476 22.14781 21.78887 22.36290 22.11394
#> [340,] 22.52962 22.76685 22.68607 22.64264 22.58256 22.28768
#> [341,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [342,]      NaN      NaN      NaN      NaN 22.58575      NaN
#> [343,] 25.72840 25.22640 25.85966 25.77985 25.69984 25.20384
#> [344,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [345,] 25.99475 24.86491 25.04802 24.81622 25.56581 24.90100
#> [346,] 21.72692 21.82194 22.14193 21.94806 22.43858 21.58417
#> [347,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [348,] 22.59825 22.48598 22.43593 22.63742 22.47321 22.45297
#> [349,]      NaN      NaN      NaN      NaN 20.82446      NaN
#> [350,] 23.88418 24.05482 24.17356 23.99542 23.97071 24.08359
#> [351,]      NaN      NaN      NaN      NaN 25.67050      NaN
#> [352,]      NaN 21.85544 21.31002 21.25303 21.19671 20.73157
#> [353,] 27.77203 27.82098 27.93989 27.90472 27.83304 28.02189
#> [354,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [355,]      NaN      NaN      NaN 24.83937      NaN 24.71717
#> [356,] 24.28390 24.16300 24.25515 25.74146 25.57901 25.56580
#> [357,]      NaN      NaN 23.10613      NaN      NaN      NaN
#> [358,] 26.60608 26.46775 26.33528 26.44903 25.58187 26.54660
#> [359,] 20.39770 18.48801 23.45714 20.19099 19.30855 19.33418
#> [360,] 25.12022 24.84879 25.07578 25.27699 25.18232 25.13864
#> [361,] 22.12054 21.78756 21.56946      NaN 22.09326 21.72729
#> [362,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [363,] 23.00823 22.62945 21.91148 23.39767 23.24056 23.20628
#> [364,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [365,] 24.89524 24.85507 25.19328 24.95667 24.92811 24.97008
#> [366,]      NaN      NaN 26.34562      NaN      NaN      NaN
#> [367,] 22.13821      NaN 22.86361 21.53469 21.62565 22.24013
#> [368,] 23.47717 23.82377 23.25407 23.91242 23.99904 23.81548
#> [369,] 23.16359 23.11291 23.30520 23.34844 23.02986 23.08483
#> [370,] 25.55492 25.30002 25.58081 25.55603 25.37796 25.47745
#> [371,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [372,] 22.69662 22.04597      NaN 22.18634 22.59571      NaN
#> [373,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [374,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [375,] 24.21208 23.88176 23.88837 24.22361 23.85091 23.68560
#> [376,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [377,] 25.77145 25.70754 25.82556 25.38805 25.31883 25.32804
#> [378,] 24.24205 23.27952 24.30020 23.57704 23.74735      NaN
#> [379,] 25.68752 25.45729 25.52076 25.98553 25.45563 25.62731
#> [380,] 25.23473 25.57822 25.38281 25.36993 25.33565 25.66348
#> [381,] 24.92799 24.74032 26.00866 25.04801 25.06069 25.03209
#> [382,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [383,] 25.06050 25.09999 25.14361 24.90180 24.89284 24.98741
#> [384,] 22.12174      NaN 23.91397 22.10002      NaN      NaN
#> [385,] 24.83848 25.11735 25.22222 25.43303 25.28591 25.51060
#> [386,] 24.46176 24.34306 24.24680 24.60346 24.51476 24.58116
#> [387,] 22.97004 22.41529 22.68136      NaN 22.27870 22.93535
#> [388,] 24.11862 23.92370 23.87089 24.06011 24.32320 23.83682
#> [389,] 26.53308 26.58258 26.37237 27.31424 27.45278 27.28295
#> [390,] 23.98494 23.90702 23.66936 23.55356 23.92162 24.72088
#> [391,] 22.63283 22.48013 22.54741 22.54875 22.86109 22.70793
#> [392,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [393,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [394,] 24.85610 24.52761 24.83186 24.43212 24.42124 24.45507
#> [395,] 27.80748 27.63539 28.22894 27.84250 27.64492 27.95104
#> [396,] 24.56471 23.97299 23.34706 24.47772 24.25818 24.52001
#> [397,] 22.27631      NaN      NaN      NaN      NaN      NaN
#> [398,] 27.68326 27.59625 27.60646 27.87364 27.67584 27.80907
#> [399,] 24.70906 24.69080 24.80509 24.70924 24.65182 24.61232
#> [400,]      NaN 23.49179 24.19626      NaN 24.06661 23.24489
#> [401,] 24.75562 24.43919 24.60944 24.29503 24.14725 24.50630
#> [402,] 23.59708 23.41983 23.73014 23.49875 23.39231      NaN
#> [403,] 24.91123 25.02251 25.33409 25.28380 25.16875 25.20676
#> [404,] 26.30715 23.90329 25.48997 26.13024 24.17078 24.70641
#> [405,]      NaN 23.15795      NaN      NaN 22.37215      NaN
#> [406,] 24.34799 23.97111 24.21656 24.23234 24.28876 24.05366
#> [407,] 22.29160 22.40148 22.26152 21.83927 21.97470 22.20979
#> [408,]      NaN      NaN 23.18189      NaN      NaN      NaN
#> [409,] 25.72109 25.84338 24.97351 24.96899 24.88600 24.82789
#> [410,] 30.60525 30.56843 30.67496 30.60526 30.54657 30.63189
#> [411,] 22.45361 23.08952 22.73981 23.84556 23.06984 22.56653
#> [412,] 25.81473 24.52391 24.49169 24.34722 24.26482 24.42556
#> [413,] 23.34055 23.15687 22.86378 22.66733 24.69397 22.94571
#> [414,] 25.30424 25.43492 25.22122 25.85524 25.92728 25.85305
#> [415,] 24.74581 24.47067 24.24793 24.47450 24.35108 24.27285
#> [416,] 27.61161 27.72618 27.75812 27.98577 27.91274 27.99394
#> [417,]      NaN      NaN 24.13868      NaN      NaN      NaN
#> [418,] 20.66397 19.93891 23.88679 20.41338 20.08447 20.51817
#> [419,] 27.57273 27.53932 27.56215 27.67759 27.67170 27.71252
#> [420,] 28.93160 28.91234 28.89891 28.78244 28.78803 28.80378
#> [421,] 29.99415 30.02789 29.95199 30.12805 30.04752 30.04583
#> [422,] 27.58118 27.53823 27.54141 27.68425 27.85781 27.76066
#> [423,] 28.50436 28.51719 28.61062 28.68510 28.65239 28.79915
#> [424,] 22.11496 22.73248 22.44670 23.07584 22.83902 23.66460
#> [425,] 30.07839 30.12386 30.10525 30.07597 30.05414 30.13418
#> [426,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [427,]      NaN 26.99540 26.96741 25.14706 25.18391 24.65228
#> [428,] 26.80772 26.63782 26.81946 26.61386 26.68886 26.45366
#> [429,] 26.16682 26.21666 26.13034 25.87975 26.00862 25.96532
#> [430,] 27.38639 27.35821 27.42925 27.44400 27.43828 27.38206
#> [431,] 23.50543 23.51135 23.77364 23.39205 23.02501 23.39545
#> [432,] 26.73637 26.72399 26.81294 26.70829 26.72507 26.72420
#> [433,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [434,] 26.66925 26.74448 26.84954 28.05748 28.07304 28.14861
#> [435,] 29.54392 29.62803 29.69031 29.62357 29.57183 29.49797
#> [436,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [437,] 20.47892      NaN 20.82765 20.24516 20.03389 20.66778
#> [438,] 25.01647 25.16567 25.06046 24.98932 24.80754 24.91487
#> [439,]      NaN      NaN 21.85901      NaN      NaN      NaN
#> [440,] 27.34824 27.30610 27.31291 25.95257 24.09811 26.31617
#> [441,] 23.36324 24.29138 23.87356 23.66579 24.05104 23.89960
#> [442,] 25.58003 25.63876 25.51307 25.48422 25.45767 25.60278
#> [443,] 25.98618 25.91796 26.16521 25.85391 25.83971 26.05550
#> [444,] 29.27895 29.28217 29.27935 29.34506 29.28401 29.27513
#> [445,]      NaN      NaN 23.43161      NaN      NaN 20.02770
#> [446,]      NaN 22.29735      NaN      NaN      NaN 21.85939
#> [447,] 21.63817 21.65718 21.14737 21.46732 21.81948 21.49267
#> [448,]      NaN      NaN      NaN 23.92080 23.85529 23.76860
#> [449,] 27.25432 27.20634 27.13883 25.61790 25.72944 25.72002
#> [450,] 23.06865 22.69128 22.21006 23.81059 22.73857 22.43357
#> [451,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [452,] 25.84372 25.80351 25.93845 25.72557 25.82995 25.87575
#> [453,] 25.34371 25.28729 25.15969 25.54557 25.48919 25.22935
#> [454,] 21.38190 21.97320 21.62725 22.33418 21.66830 21.55183
#> [455,] 24.13837 24.04368 24.55988 24.25609 24.31465 23.90702
#> [456,] 24.37458 24.63198 24.43369 24.16556 24.36379 24.18314
#> [457,] 27.53905 27.36194 27.58484 25.89715 25.87368 26.00528
#> [458,] 26.09894 27.33570 27.14561 26.04608 25.92568 26.57951
#> [459,] 24.34476 24.24474 24.47110      NaN 22.79066 22.70586
#> [460,] 24.58932 24.03305 24.26635 24.16676 23.67422 24.06373
#> [461,] 24.79650 24.66275 24.94525 24.91994 24.80085 24.89481
#> [462,] 23.83820 24.03584 24.67641 24.08008 24.34345 23.84003
#> [463,] 26.39277 26.44361 26.33269 25.15503 25.28574 25.15623
#> [464,]      NaN      NaN      NaN      NaN      NaN 22.41687
#> [465,] 26.82845 26.90127 26.70213 26.77178 26.73982 26.76737
#> [466,] 27.19667 27.17579 26.74248 26.49271 26.58035 26.48924
#> [467,]      NaN      NaN      NaN 24.08788 23.86590 24.33490
#> [468,] 23.65938 23.92319 24.42911 23.39599 24.17614 23.75105
#> [469,] 23.56256 23.55368 23.61885 23.43924 23.47358 23.51485
#> [470,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [471,] 26.31339 26.23874 26.51721 26.13938 26.28919 26.29605
#> [472,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [473,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [474,] 24.32288 24.22551 24.57528 24.19929 24.32838 24.41473
#> [475,] 26.48442 26.41387 26.48235 24.91552 24.94437 24.90940
#> [476,] 26.59149 26.16486 26.48167 26.78203 26.81052 26.79870
#> [477,] 23.69891 23.28701 23.98598 23.79551 23.64647 23.45375
#> [478,] 24.82163 24.87666 25.01495 24.76502 24.73433 24.95584
#> [479,] 24.99000 25.38196 25.65630 25.63495 25.48798 25.53827
#> [480,] 24.51281 24.68849 25.01519 25.08055 24.92715 25.20664
#> [481,] 23.64845 23.48431 23.75827 24.42803 24.10254 24.14957
#> [482,] 22.70316 21.53872 24.29411 22.72013 22.11569 23.14995
#> [483,] 25.71683 26.99146 27.11534 25.55631 25.52412 25.55875
#> [484,] 24.73486 25.39767 24.68992 25.37859 25.18575 25.30103
#> [485,] 27.76541 27.64966 27.80199 25.92225 26.00477 26.05329
#> [486,] 22.93276 22.90967 23.68902      NaN 22.99762      NaN
#> [487,] 23.81158 25.65736 25.08691 24.00707 25.64251 25.71905
#> [488,]      NaN 22.31363 22.33304      NaN 22.31269 22.82153
#> [489,] 25.55584 25.56651 25.64290 25.56145 25.50479 25.46314
#> [490,] 23.76385 23.74796 23.80282 23.83673 23.50809 23.95260
#> [491,] 26.49109 26.35642 26.40695 25.80869 25.62525 25.94446
#> [492,] 27.11316 27.12303 27.18585 27.08698 27.10924 27.17884
#> [493,] 25.59525 25.20617 25.53961 24.38001      NaN 24.20851
#> [494,]      NaN      NaN      NaN 22.30025 21.96856 22.26682
#> [495,] 26.30008 26.67248 26.55009 26.74558 26.98227 27.11330
#> [496,] 22.14034 24.02056 24.11545 23.53955 24.11156 23.75297
#> [497,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [498,] 23.27997 22.41910 22.55096 23.67422 22.75602 23.18798
#> [499,] 28.50904 28.51683 28.81094 28.59124 28.52798 28.58803
#> [500,]      NaN 23.48665 23.90665      NaN      NaN 23.47172
#> [501,] 24.25875 24.12542 24.03322 24.20538 24.08829 24.28228
#> [502,] 20.59177 20.91947      NaN 22.46660 22.92962 22.31327
#> [503,] 25.91775 25.58892 25.79333 25.58603 25.79077 25.76709
#> [504,] 23.37186 23.34746 23.10096      NaN 23.19409 22.99258
#> [505,] 24.46220 24.91848 24.61756 21.46881 24.41203 21.88662
#> [506,] 22.63034 22.79029 22.76937 23.15689 22.94473 23.36866
#> [507,] 22.22935 22.76119      NaN      NaN 21.94614 21.92864
#> [508,] 22.34188 22.52884      NaN 23.31714      NaN      NaN
#> [509,]      NaN      NaN      NaN      NaN 23.68271      NaN
#> [510,]      NaN      NaN      NaN 25.55678 25.44862 25.60546
#> [511,] 24.56703 24.34672 24.43192 24.33429 24.35474 24.58526
#> [512,] 23.24124      NaN 23.04365 22.88703 23.45739      NaN
#> [513,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [514,] 22.78546 22.66917      NaN 22.63653 22.98184 23.32251
#> [515,]      NaN 22.37645 22.92770 22.71121 22.81204      NaN
#> [516,] 26.08746 25.87825 26.07794 25.95624 25.78452 25.52877
#> [517,] 21.78285 21.72684 21.87579 22.28608 22.18435 21.92626
#> [518,] 25.34245 25.40022 25.49531 25.73348 25.35736 25.72952
#> [519,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [520,] 24.29168 24.39473 24.51280 24.48614 24.50642 24.50237
#> [521,] 23.86212 23.83817 23.78177 24.14048 23.69349 23.86372
#> [522,] 24.24972 24.01859 24.41244 24.31763 24.76344 24.67126
#> [523,] 28.49137 28.49812 28.61709 28.79634 28.65174 28.67125
#> [524,] 27.00346 27.06979 25.75521 23.96477 27.04781 26.99518
#> [525,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [526,] 22.72648 22.85719 23.14233 23.11294 23.12917 23.18982
#> [527,] 24.19016 24.23557 24.31534 24.50416 24.25746 24.21941
#> [528,] 23.71240 23.02159 21.65263      NaN      NaN 21.85133
#> [529,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [530,] 26.59206 26.73467 26.71986 26.61344 26.61583 26.63686
#> [531,] 22.82180 23.68849 23.82250 23.67874 23.25312 23.50458
#> [532,] 26.12814 25.95136 26.33422 26.60215 26.34761 26.03046
#> [533,] 25.63145 25.73154 25.99878 25.72070 25.78753 25.74031
#> [534,] 24.21963 24.19731 24.39434 24.30703 24.13226 24.45870
#> [535,] 23.76375      NaN 23.73738 23.82367 24.08383 24.22729
#> [536,] 22.97836 22.41356 22.41746 22.90005 22.57316 22.83959
#> [537,] 30.15774 29.97856 30.04934 29.96072 29.84721 29.86616
#> [538,]      NaN 19.97561      NaN      NaN      NaN      NaN
#> [539,] 24.89042 25.09455 25.10778 25.20017 24.92217 25.01588
#> [540,] 26.64113 26.44774 26.66221 25.03741 24.81142 25.02187
#> [541,]      NaN      NaN      NaN 23.98564 24.00230 24.00445
#> [542,] 26.26151 26.16710 26.47859 24.51455 24.57952 24.83880
#> [543,] 24.33617 24.56510 24.74759 24.03357 24.49086 24.65528
#> [544,]      NaN      NaN 23.57219      NaN 22.72084 22.19764
#> [545,]      NaN 24.20352 24.09434 24.39172 23.74314 23.73531
#> [546,]      NaN      NaN      NaN      NaN      NaN 25.97520
#> [547,] 27.42502 27.04718 24.67196      NaN      NaN      NaN
#> [548,] 25.71319 25.78004 26.08930 26.02471 25.98839 26.26064
#> [549,] 24.16154 24.02030 24.16461 24.21119 24.21549 24.09313
#> [550,] 21.97376 20.95156 23.60399 21.84518 22.41160 22.98021
#> [551,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [552,]      NaN 22.26965 23.02704 22.50368 23.18082 23.09241
#> [553,]      NaN      NaN 22.04560 22.60080 22.90277      NaN
#> [554,] 26.27768 26.30019 26.42598 26.59036 26.59958 26.69659
#> [555,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [556,] 22.65279 22.79082 22.68464      NaN 22.89049 23.29081
#> [557,] 24.15430 24.11648 23.71649      NaN 23.83161 24.26964
#> [558,] 24.50549 23.41635 24.14903 22.77981 23.10484 23.35380
#> [559,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [560,] 27.54959 27.36936 27.40649 27.51510 27.51803 27.64870
#> [561,] 28.77266 28.61295 28.80104 28.71967 28.65578 28.80464
#> [562,] 27.13626 27.16054 27.22204 27.21742 27.35515 27.41344
#> [563,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [564,] 26.31180 26.43598 26.77077 26.57421 26.51847 26.41766
#> [565,]      NaN      NaN      NaN 24.30973 26.15848 26.46449
#> [566,] 23.21328 23.52289 23.20629 22.45066 22.33137 23.52981
#> [567,] 24.50021 24.22007 24.16500      NaN      NaN      NaN
#> [568,] 27.88105 27.76154 28.00860 27.78019 27.80520 27.86673
#> [569,]      NaN      NaN      NaN      NaN 21.94764      NaN
#> [570,]      NaN      NaN 26.36378      NaN      NaN      NaN
#> [571,]      NaN      NaN      NaN      NaN 22.47823      NaN
#> [572,] 24.42188 23.56116      NaN 24.77535 23.77545      NaN
#> [573,] 24.90559 25.08869 25.08966 25.22718 25.22273 25.47568
#> [574,] 24.85472 24.56837 24.79724 24.79645 24.87886 24.76319
#> [575,] 22.35361      NaN      NaN 22.10875      NaN      NaN
#> [576,] 21.81722      NaN 22.16791 21.94567 21.95745 21.52483
#> [577,] 24.56331 24.60524 24.71277 24.41590 24.43460 24.42368
#> [578,] 27.62415 27.47738 27.51871 27.51171 27.52006 27.54067
#> [579,] 24.43441 24.02957 23.79098 24.79351 24.54099 25.78614
#> [580,] 23.55532 23.68250 23.17019 23.42278 23.68902      NaN
#> [581,] 28.26154 28.20695 28.20158 28.76459 28.45846 28.54019
#> [582,] 24.96146 24.98846 24.90495 25.02064 24.79862 25.13809
#> [583,] 25.88010 25.66955 25.68416 25.74448 25.75768 25.78374
#> [584,] 22.68742 22.65762 22.80692 22.76634 22.41083 22.82382
#> [585,] 28.59303 28.24098 28.50348 28.66988 28.57942 28.29656
#> [586,] 25.58003 25.63876 25.51307 25.48422 25.45767 25.60278
#> [587,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [588,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [589,] 24.32649      NaN 24.37915 24.15908      NaN      NaN
#> [590,] 24.99524 25.24205 25.09035 25.09483 25.24402 25.35326
#> [591,] 23.49668 23.35582 23.00946 23.87652 23.80518 23.72557
#> [592,] 24.28821 24.20933 24.35084 24.18511 22.92608 24.00740
#> [593,] 24.05499 24.26607 24.50057 24.64052 24.56720 24.48796
#> [594,] 26.66017 25.36122 27.19396 26.78173 25.23153 27.13501
#> [595,]      NaN 24.75944      NaN 24.64653      NaN      NaN
#> [596,]      NaN 22.78445 23.23149 19.64816      NaN 22.90064
#> [597,] 23.94842 24.05084 24.25039 24.07725 23.38034 24.01648
#> [598,]      NaN      NaN 24.34082 23.53891 23.31846 23.79263
#> [599,] 23.40587 23.14043 22.96955 22.96886 23.11694 23.18929
#> [600,] 24.01725 23.97106 24.07022 24.11783 23.95180 24.11696
#> [601,] 23.30020 23.35111 23.17295 24.01461      NaN      NaN
#> [602,] 23.95571 23.92696 24.06143 23.66046 24.19460 23.99404
#> [603,] 24.76289 26.13266 27.12278 24.94138 26.26276 26.88960
#> [604,] 24.38265 24.05408 24.57277 24.05573 24.05961 24.11854
#> [605,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [606,] 24.88409 24.57617 24.75567 24.68024 24.67589 24.86825
#> [607,] 25.62672 25.59114 25.40019 25.05254 25.18956 25.24532
#> [608,]      NaN      NaN      NaN 27.86118 27.87426 27.88373
#> [609,] 22.56639 22.07777 23.36346 22.28932 22.04303 22.59234
#> [610,] 26.28854 26.19549 26.50259 26.13475 26.23692 26.28198
#> [611,] 26.41578 26.41257 26.59643 26.71145 26.78221 26.70398
#> [612,] 22.85356 22.89565 23.49412 23.23786 22.84084 23.29866
#> [613,]      NaN      NaN      NaN 25.82019 24.66500 25.72033
#> [614,] 22.40932 23.02910      NaN      NaN      NaN      NaN
#> [615,] 23.29250 23.41914 23.89391 23.20695 23.09355 23.15593
#> [616,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [617,] 22.68571 23.13232 22.46835 22.72178 22.68638 22.36136
#> [618,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [619,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [620,] 24.82357 25.03149 24.97561 25.29719 25.05345 25.36913
#> [621,] 23.81176 24.08001 23.93715 23.92252 23.96710 23.47420
#> [622,] 22.63973 23.21061 22.92646 22.82770 22.80359 23.62723
#> [623,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [624,] 23.18826 22.74154 22.94633      NaN      NaN      NaN
#> [625,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [626,] 24.06669 24.16023 23.96622      NaN      NaN      NaN
#> [627,] 24.65436 24.67093 25.04843 24.74335 24.64818 24.91159
#> [628,] 23.35340 23.05892 23.33033 23.60082 23.72920 23.88883
#> [629,]      NaN      NaN      NaN 25.16181 25.23623 25.74643
#> [630,] 24.63151 24.29985 24.68624 24.41680 24.41777 24.41622
#> [631,] 24.64719 24.63544 25.02914 24.21395 24.10604 24.13541
#> [632,] 24.84067 24.92134 24.63290 25.03149 24.70150 24.39571
#> [633,] 22.49831      NaN      NaN 22.53360 22.34335 22.64768
#> [634,] 25.15804      NaN 24.99265 25.02758 24.95925 24.92945
#> [635,]      NaN      NaN      NaN 25.05308 24.73076 25.13097
#> [636,] 23.56139 23.53623 23.39113 23.15508 23.44721 23.34962
#> [637,]      NaN      NaN 24.02741 23.22609 24.18775 23.89013
#> [638,] 24.43059 24.11744      NaN 23.86938 23.34340      NaN
#> [639,] 22.20108 22.00361 22.62825 22.17192 22.36242 22.17864
#> [640,]      NaN      NaN 23.03110      NaN      NaN      NaN
#> [641,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [642,] 23.80658 23.25689 24.01249 23.66037 23.67461 23.96424
#> [643,] 20.83499      NaN      NaN 21.24010 21.68132      NaN
#> [644,]      NaN      NaN      NaN      NaN 22.28045      NaN
#> [645,] 26.56522 24.95505 22.84824 25.98203 25.54424 25.31403
#> [646,] 25.25494 25.16158 25.25868 25.05858 25.28584 25.20747
#> [647,]      NaN      NaN      NaN      NaN 21.65927      NaN
#> [648,] 28.42174 27.18565 28.36123 27.25746 27.21317 27.22527
#> [649,] 27.14948 27.09386 27.24227 27.04041 26.92162 26.95105
#> [650,] 24.53961 25.59315 25.47011 25.58712 25.54242 25.58124
#> [651,] 24.12131 24.25054 24.44872      NaN      NaN      NaN
#> [652,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [653,] 26.54501 26.04951 27.03685 26.42466 26.31855 26.59263
#> [654,] 22.99452 22.53541 22.49599 22.49740 22.06299 22.40219
#> [655,] 24.16454 24.36426 24.16362 23.39008 23.48702 24.12384
#> [656,] 22.19877 22.25710 22.59621 22.49064 22.72711 22.44821
#> [657,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [658,] 29.00603 28.83993 29.03708 28.87150 28.83957 28.93340
#> [659,] 23.59744 22.80168 23.60366 23.41510 23.57007 23.70258
#> [660,] 28.24617 28.13922 28.13482 28.08977 28.01855 28.06317
#> [661,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [662,]      NaN      NaN      NaN      NaN      NaN      NaN
#> [663,] 22.73105 22.69967 23.08737 23.01793 22.76128 22.87496
#> [664,] 25.91857 25.88944 26.12451 26.11957 25.95360 25.90442
#> [665,] 24.15414 24.05250 24.43848 24.55368 24.64037 24.80196
#> [666,]      NaN      NaN 21.08213      NaN      NaN      NaN
#> [667,]      NaN 23.74848      NaN      NaN      NaN 23.47346
#> [668,] 24.95549 24.69938 25.32684 24.47373 24.18887 24.48282
#> [669,] 24.21897 24.25775 24.17265 24.01899 23.76496 24.06898
#> [670,] 27.79347 27.71151 27.80093 28.05850 27.76337 27.79447